% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{autoplot.dna_barplot}
\alias{autoplot.dna_barplot}
\title{Plot \code{dna_barplot} object.}
\usage{
\method{autoplot}{dna_barplot}(
  object,
  ...,
  lab.pos = "Agreement",
  lab.neg = "Disagreement",
  lab = TRUE,
  colors = FALSE,
  fontSize = 12,
  barWidth = 0.6,
  axisWidth = 1.5,
  truncate = 40,
  exclude.min = NULL
)
}
\arguments{
\item{object}{A \code{dna_barplot} object.}

\item{...}{Additional arguments; currently not in use.}

\item{lab.pos, lab.neg}{Names for (dis-)agreement labels.}

\item{lab}{Should (dis-)agreement labels and title be displayed?}

\item{colors}{If \code{TRUE}, the \code{Colors} column in the
\code{dna_barplot} object will be used to fill the bars. Also accepts
character objects matching one of the attribute variables of the
\code{dna_barplot} object.}

\item{fontSize}{Text size in pt.}

\item{barWidth}{Thickness of the bars. Bars will touch when set to \code{1}.
When set to \code{0.5}, space between two bars is the same as thickness of
bars.}

\item{axisWidth}{Thickness of the x-axis which separates agreement from
disagreement.}

\item{truncate}{Sets the number of characters to which axis labels should be
truncated.}

\item{exclude.min}{Reduces the plot to entities with a minimum frequency of
statements.}
}
\description{
Plot a barplot generated from \code{\link{dna_barplot}}.
}
\details{
This function plots \code{dna_barplot} objects generated by the
\code{\link{dna_barplot}} function. It plots agreement and disagreement with
DNA statements for different entities such as \code{"concept"},
\code{"organization"}, or \code{"person"}. Colors can be modified before
plotting (see examples).
}
\examples{
\dontrun{
dna_init()
dna_sample()

dna_openDatabase("sample.dna", coderId = 1, coderPassword = "sample")

# compute barplot data
b <- dna_barplot(statementType = "DNA Statement",
                 variable = "concept",
                 qualifier = "agreement")

# plot barplot with ggplot2
library("ggplot2")
autoplot(b)

# use entity colours (here: colors of organizations as an illustration)
b <- dna_barplot(statementType = "DNA Statement",
                 variable = "organization",
                 qualifier = "agreement")
autoplot(b, colors = TRUE)

# edit the colors before plotting
b$Color[b$Type == "NGO"] <- "red"         # change NGO color to red
b$Color[b$Type == "Government"] <- "blue" # change government color to blue
autoplot(b, colors = TRUE)

# use an attribute, such as type, to color the bars
autoplot(b, colors = "Type") +
  scale_colour_manual(values = "black")

# replace colors for the three possible actor types with custom colors
autoplot(b, colors = "Type") +
  scale_fill_manual(values = c("red", "blue", "green")) +
  scale_colour_manual(values = "black")
}

}
\seealso{
Other {rDNA barplots}: 
\code{\link{dna_barplot}()},
\code{\link{print.dna_barplot}()}
}
\author{
Johannes B. Gruber, Tim Henrichsen
}
\concept{{rDNA barplots}}
