% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_addDocument}
\alias{dna_addDocument}
\title{Add a document to the DNA database}
\usage{
dna_addDocument(connection, title = "", text = "", coder = 1,
  author = "", source = "", section = "", notes = "", type = "",
  date = Sys.time(), returnID = FALSE, verbose = TRUE)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{title}{A character object containing the title of the new document.}

\item{text}{A character object containing the text of the new document. Line
breaks can be included as \code{"\\n"}.}

\item{coder}{An integer value indicating which coder created the document.}

\item{author}{A character object containing the author of the document.}

\item{source}{A character object containing the source of the document.}

\item{section}{A character object containing the section of the document.}

\item{notes}{A character object containing notes about the document.}

\item{type}{A character object containing the type of the document.}

\item{date}{A \code{POSIXct} object containing the date/time stamp of the
document. Alternatively, the date/time can be supplied as an integer
value indicating the milliseconds since the start of 1970-01-01.}

\item{returnID}{Return the ID of the newly created document as a numeric
value?}

\item{verbose}{Print details?}
}
\description{
Add a new document to the DNA database.
}
\details{
The \code{dna_addDocument} function can add new documents to an existing DNA
database. The user supplies a \link{dna_connection} object as well as various
details about the document, for example the title, text, date etc. The date
can be supplied either as a \link{POSIXct} object or as an integer value
containing millisecond since 1970-01-01. The document ID will be
automatically generated and can be returned if \code{returnID} is set to
\code{TRUE}.
}
\author{
Philip Leifeld
}
