% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_addDocuments}
\alias{dna_addDocuments}
\title{Add one or more documents to the DNA database}
\usage{
dna_addDocuments(
  coder_id = 1,
  title = "",
  text = "",
  author = "",
  source = "",
  section = "",
  type = "",
  notes = "",
  date_time = Sys.time()
)
}
\arguments{
\item{coder_id}{An integer value indicating which coder creates the document.}

\item{title}{The title(s) of the new document(s).}

\item{text}{The text(s) of the new document(s).}

\item{author}{The author(s) of the new document(s).}

\item{source}{The source(s) of the new document(s).}

\item{section}{The section(s) of the new document(s).}

\item{type}{The type(s) of the new document(s).}

\item{notes}{The notes of the new document(s).}

\item{date_time}{The date(s)/time(s) of the new document(s). Provided as a
vector (potentially of length 1 if only one document is added) of either
numeric objects or POSIXct objects. POSIXct objects naturally store full
date/time information. Numeric objects expect date/time to be indicated as
epoch seconds since 1 January 1970 UTC.}
}
\value{
The document ID(s) of the newly created document(s) in the database.
}
\description{
Add one or more documents to the DNA database.
}
\details{
The \code{dna_addDocuments} function can add new documents to an existing
DNA database. The user supplies a single coder ID for the coder to whom the
new documents should belong and several equally long vectors of information
for the titles, texts, authors, dates/times, etc. of the new documents. To
add a single document, these vectors can have length 1. The dates/times must
be provided as a vector of POSIXct objects or a vector of numeric objects
indicating epoch seconds since 1 January 1970 UTC. After adding the documents
to the database, the function returns the generated document IDs.
}
\seealso{
Other documents: 
\code{\link{dna_deleteDocuments}()},
\code{\link{dna_getDocuments}()},
\code{\link{print.dna_documents}()}
}
\author{
Philip Leifeld
}
\concept{documents}
