% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_addRegex}
\alias{dna_addRegex}
\title{Add a new regular expression to the database}
\usage{
dna_addRegex(connection, regex, color = "#FFFF00")
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{regex}{A regular expression.}

\item{color}{A color in the form of a hexadecimal RGB string, such as
\code{"#FFFF00"} for yellow.}
}
\description{
Add a new regular expression to the database.
}
\details{
Add a new statement type to a database. The statement type contains no
variables but can be populated with variables using the
\link{dna_addVariable} function. Along with the the label used to describe
the statement type, a color needs to be supplied in order to display the
statement type in this color in the GUI (see \link{dna_gui}).
}
\author{
Philip Leifeld
}
