% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_addStatement}
\alias{dna_addStatement}
\title{Add a statement to the DNA database}
\usage{
dna_addStatement(connection, documentID, startCaret = 1, endCaret = 2,
  statementType = "DNA Statement", coder = 1, returnID = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{documentID}{An integer specifying the ID of the document for which the
statement should be added.}

\item{startCaret}{An integer for the start location of the statement in the
document text. Must be non-negative and not larger than the number of
characters minus one in the document.}

\item{endCaret}{An integer for the stop location of the statement in the
document text. Must be non-negative, greater than \code{startCaret}, and not
larger than the number of characters in the document.}

\item{statementType}{The statement type of the statement that will be added.
Can be provided as an integer ID of the statement type or as a character
object representing the name of the statement type (if there is no
ambiguity).}

\item{coder}{An integer value indicating which coder created the document.}

\item{returnID}{Return the ID of the newly created statement as a numeric
value?}

\item{verbose}{Print details?}

\item{...}{Values of the variables contained in the statement, for example
\code{organization = "some actor", concept = "my concept", agreement = 1}.
Values for Boolean variables can be provided as \code{logical} values
(\code{TRUE} or \code{FALSE}) or \code{numeric} values (\code{1} or
\code{0}).}
}
\description{
Add a new statement to the DNA database.
}
\details{
The \code{dna_addStatement} function can add a new statement to an existing
DNA database. The user supplies a \link{dna_connection} object as well as
the document ID, location of the statement in the document, and the variables
and their values. As different statement types have different variables, the
\code{...} argument catches all variables and their values supplied by the
user. The statement ID will be automatically generated and can be returned
if \code{returnID} is set to \code{TRUE}.
}
\author{
Philip Leifeld
}
