% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_addVariable}
\alias{dna_addVariable}
\title{Add a new variable to a statement type in the database}
\usage{
dna_addVariable(connection, statementType = 1, variable,
  dataType = "short text", simulate, verbose)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{statementType}{The statement type in which the new variable should be
defined. The statement type can be supplied as an integer ID or character
string, for example \code{1} or \code{"DNA Statement"}.}

\item{variable}{The name of the new variable as a character object. Only
characters and numbers are allowed, i.e., no whitespace characters.}

\item{dataType}{The data type of the new variable. Valid values are 
\code{"short text"} (for things like persons, organizations, locations
etc., up to 200 characters), \code{"long text"} (for things like notes,
can store more than 200 characters), \code{"boolean"} (for qualifier
variables such as a binary agreement variable), and \code{"integer"} (for
ordinal Likert scales, such as -5 to +5 or -1 to +1).}

\item{simulate}{Should the changes only be simulated instead of actually
applied to the DNA connection and the SQL database? This can help to
plan more complex recode operations.}

\item{verbose}{Print details about the recode operations?}
}
\description{
Add a new variable to a statement type in the database.
}
\details{
Add a new variable to an existing statement type in the database, based on
the statement type ID or label, a name for the new variable, and a data type
specification.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())
dna_addVariable(conn, 1, "location", "short text")
}

}
