% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_api}
\alias{dna_api}
\title{Get a reference to the headless Java class for R (API)}
\usage{
dna_api()
}
\value{
A Java object reference to the \code{Dna/HeadlessDna} class.
}
\description{
Get a reference to the headless Java class for R (API).
}
\details{
This function returns a Java object reference to the instance of the
\code{Dna/HeadlessDna} class in the DNA JAR file that is held in the rDNA
package environment and used by the functions in the package to exchange data
with the Java application. You can use the \pkg{rJava} package to access the
available functions in this class directly. API access requires detailed
knowledge of the DNA JAR classes and functions and is recommended for
developers and advanced users only.
}
\examples{
\dontrun{
library("rJava") # load rJava package to use functions in the Java API
dna_init()
dna_sample()
dna_openDatabase(coderId = 1,
                 coderPassword = "sample",
                 db_url = "sample.dna")
api <- dna_api()

# use the \code{getVariables} function to retrieve variables
variable_references <- api$getVariables("DNA Statement")

# iterate through variable references and print their data type
for (i in seq(variable_references$size()) - 1) {
  print(variable_references$get(as.integer(i))$getDataType())
}
}

}
\seealso{
Other database: 
\code{\link{dna_closeDatabase}()},
\code{\link{dna_openConnectionProfile}()},
\code{\link{dna_openDatabase}()},
\code{\link{dna_printDetails}()},
\code{\link{dna_saveConnectionProfile}()}
}
\author{
Philip Leifeld
}
\concept{database}
