% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_attributes}
\alias{dna_attributes}
\title{Retrieve attributes from DNA's attribute manager}
\usage{
dna_attributes(connection, statementType = "DNA Statement",
  variable = "organization")
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the 
\code{dna_connection} function.}

\item{statementType}{The name of the statement type in which the variable 
of interest is nested. For example, \code{"DNA Statement"}.}

\item{variable}{The variable for which the attributes should be retrieved. 
For example, \code{"organization"} or \code{"concept"}.}
}
\description{
Retrieve attributes for a given statement type and variable.
}
\details{
For any variable in DNA (e.g., organization or concept), the attribute 
manager in DNA stores four attributes, or meta-variables, for each value. 
For example, an organization can be associated with a color, a type, 
an alias, and notes. The \code{dna_attributes} function serves to retrieve 
these attributes for all values of a given variable from DNA. To be sure 
that the attributes are retrieved for the right variable, the user also 
has to provide the statement type in which the variable is nested. Note that 
the \code{dna_attributes} function returns the attributes for all values of 
the respective variable (similar to the \code{includeIsolates = TRUE} 
argument of the \code{dna_network} function. The attributes are returned as 
a data frame with the ID of the value as the first column, the actual value 
as the second column, the color as the third column, type as the fourth 
column, alias as fifth, notes as sixth, and the number of occurrences of the 
value as the seventh column. The ninth column indicates whether the value 
is used anywhere in the dataset, and the tenth and last column indicates if 
the value is contained in the current connection.
}
\examples{
download.file("https://github.com/leifeld/dna/releases/download/v2.0-beta.19/dna-2.0-beta19.jar", 
              destfile = "dna-2.0-beta19.jar", mode = "wb")
download.file("https://github.com/leifeld/dna/releases/download/v2.0-beta.19/sample.dna", 
              destfile = "sample.dna", mode = "wb")
dna_init("dna-2.0-beta19.jar")
conn <- dna_connection("sample.dna")
at <- dna_attributes(conn, "DNA Statement", "organization")
at
}
