% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_cluster}
\alias{dna_cluster}
\title{Cluster network from a DNA connection}
\usage{
dna_cluster(connection, variable = "organization", rm.duplicates = FALSE,
  clust.method = "ward.D2", qualifier = "agreement", colours = "color",
  cut.k = NULL, cut.h = NULL, ...)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\link{dna_connection} function.}

\item{variable}{The first variable for network construction  (see
\link{dna_network}). The second one defaults to "concept" but can be
provided via ... if necessary.}

\item{rm.duplicates}{Should duplicates be removed from documents in the
database (see \link{dna_network}. This also influences if
\link[vegan]{vegdist} or \link[stats]{dist} is used to calculate the
dissimilarity matrix.}

\item{clust.method}{The agglomeration method to be used. See
\link[stats]{hclust}.}

\item{qualifier}{The qualifier variable. See \link{dna_network}.}

\item{colours}{Which attribute of variable from DNA should be used to assign
colours? Additionally, can be "membership" if cut.k or cut.h are provided.}

\item{cut.k, cut.h}{See k and h in \link[stats]{cutree}.}

\item{...}{additional arguments passed to \link{dna_network}}
}
\description{
Clustering methods for DNA connections.
}
\details{
Perform a cluster analysis based on a DNA connection.
}
\examples{
\dontrun{
dna_init("dna-2.0-beta20.jar")
conn <- dna_connection(dna_sample())

clust.l <- dna_cluster(connection)

dna_plotCluster(clust.l)
}
}
\author{
Johannes B. Gruber
}
