% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_colorStatementType}
\alias{dna_colorStatementType}
\title{Change the color of a statement type}
\usage{
dna_colorStatementType(connection, statementType, color = "#FFFF00")
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{statementType}{The statement type whose color should be updated. The
statement type can be supplied as an integer ID or character string, for
example \code{1} or \code{"DNA Statement"}.}

\item{color}{A color in the form of a hexadecimal RGB string, such as
\code{"#FFFF00"} for yellow.}
}
\description{
Change the color of a statement type.
}
\details{
This function assigns a new color to a statement type. The color is used to
display statements of this type in the DNA GUI (see \link{dna_gui}).
}
