% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_convergenceScale}
\alias{dna_convergenceScale}
\title{Convergence diagnostics for \code{dna_scale} objects}
\usage{
dna_convergenceScale(dna_scale, variable = "both", method = "geweke",
  colors = TRUE, nrow = 3, ncol = 3, value = NULL)
}
\arguments{
\item{dna_scale}{A \code{dna_scale} object.}

\item{variable}{Variable for assessing convergence diagnostics. If \code{"both"}, both
stored variables from the \code{dna_scale} function call will be used. Defaults to
\code{"both"}.}

\item{method}{Method for the convergence diagnostics. Supported are \code{"geweke"},
\code{"density"} and \code{"trace"}. Defaults to \code{"geweke"}.}

\item{colors}{Logical. Should attribute colors be used for either the
\code{density} or \code{trace} plots? Defaults to \code{TRUE}.}

\item{nrow}{Number of rows for the facet plot.}

\item{ncol}{Number of columns for the facet plot.}

\item{value}{Optional character vector if only specific actors/concepts should be
analyzed. If specified, \code{variable} will be ignored.}
}
\description{
Convergence diagnostics for the MCMC chain created by the \code{dna_scale} functions.
}
\details{
This function offers several convergence diagnostics for the MCMC chain created by the
\code{dna_scale} functions. There are three possible ways of assessing mixing of a chain:

\code{"trace"} is a graphic inspection of the sampled values by iteration. Once the chain
has reached its stationary distribution, the parameter values should look like a hairy
caterpillar, meaning that the chain should not stay in the same state for too long or
have too many consecutive steps in one direction.

\code{"density"} visually analyzes the cumulative density of the sampled values for 
each parameter. Unimodality should indicate convergence of the chain, while 
multimodality might indicate an identification problem leading to non-convergence.

\code{"geweke"} conducts a difference of means test for the sampled values for two 
sections of the chain. It then compares the first 10 percent of iterations with the final 
50 percent of iterations. If the samples are drawn from the stationary distribution of 
the chain, a difference of means test should be statistically significant at some 
conventional level (in our case 0.05). The returned test statistic is a standard Z-score. 
All values should be below the 1.96 value that would indicate significance at the 
p =< 0.05 level.
}
\examples{
\dontrun{
dna_convergenceScale(dna_scale,
                     variable = "both",
                     method = "trace",
                     colors = TRUE,
                     nrow = 4,
                     ncol = 4)

dna_convergenceScale(dna_scale,
                     method = "density",
                     colors = TRUE,
                     nrow = 1,
                     ncol = 2,
                     value = c("Senate", "Sierra Club"))
}

}
\author{
Tim Henrichsen, Johannes B. Gruber
}
