% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_deleteDocuments}
\alias{dna_deleteDocuments}
\title{Delete document(s)}
\usage{
dna_deleteDocuments(document_id)
}
\arguments{
\item{document_id}{A vector of document IDs (can be a single ID).}
}
\description{
Delete document(s) from the DNA database.
}
\details{
The \code{dna_deleteDocuments} function removes one or more documents with a
given vector of document IDs from the database.
}
\seealso{
Other documents: 
\code{\link{dna_addDocuments}()},
\code{\link{dna_getDocuments}()},
\code{\link{print.dna_documents}()}
}
\author{
Philip Leifeld
}
\concept{documents}
