% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_deleteStatements}
\alias{dna_deleteStatements}
\title{Delete statement(s)}
\usage{
dna_deleteStatements(statement_id)
}
\arguments{
\item{statement_id}{A vector of statement IDs (can be a single ID).}
}
\description{
Delete statement(s) from the DNA database.
}
\details{
The \code{dna_deleteStatements} function removes one or more statements with
a given vector of statement IDs from the database.
}
\seealso{
Other statements: 
\code{\link{dna_addStatement}()},
\code{\link{dna_getStatements}()},
\code{\link{print.dna_statements}()}
}
\author{
Philip Leifeld
}
\concept{statements}
