% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_dendrogram}
\alias{dna_dendrogram}
\title{Create a cluster dendrogram for a DNA database}
\usage{
dna_dendrogram(connection, statementType = "DNA Statement",
  variable1 = "organization", variable1Document = FALSE,
  variable2 = "concept", variable2Document = FALSE,
  qualifier = "agreement", duplicates = "include",
  start.date = "01.01.1900", stop.date = "31.12.2099",
  start.time = "00:00:00", stop.time = "23:59:59",
  excludeValues = list(), excludeAuthors = character(),
  excludeSources = character(), excludeSections = character(),
  excludeTypes = character(), invertValues = FALSE,
  invertAuthors = FALSE, invertSources = FALSE,
  invertSections = FALSE, invertTypes = FALSE, method = "best",
  k = 0, k.max = 5, rectangle.colors = NULL, labels = "value",
  label.colors = "color", label.size = 12, label.truncate = 30,
  leaf.shape = "elbow", leaf.colors = label.colors, leaf.width = 1,
  leaf.alpha = 1, symbol.shapes = 19, symbol.colors = label.colors,
  symbol.sizes = 5, circular = FALSE, theme = "bw", caption = TRUE,
  return.multiclust = FALSE)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{statementType}{The name of the statement type in which the variable
of interest is nested. For example, \code{"DNA Statement"}.}

\item{variable1}{The first variable for network construction. In a one-mode
network, this is the variable for both the rows and columns. In a
two-mode network, this is the variable for the rows only. In an event
list, this variable is only used to check for duplicates (depending on
the setting of the \code{duplicate} argument).}

\item{variable1Document}{A boolean value indicating whether the first
variable is at the document level (i.e., \code{"author"},
\code{"source"}, \code{"section"}, \code{"type"}, \code{"id"}, or
\code{"title"}).}

\item{variable2}{The second variable for network construction. In a one-mode
network, this is the variable over which the ties are created. For
example, if an organization x organization network is created, and ties
in this network indicate co-reference to a concept, then the second
variable is the \code{"concept"}. In a two-mode network, this is the
variable used for the columns of the network matrix. In an event list,
this variable is only used to check for duplicates (depending on the
setting of the \code{duplicate} argument).}

\item{variable2Document}{A boolean value indicating whether the second
variable is at the document level (i.e., \code{"author"},
\code{"source"}, \code{"section"}, \code{"type"}, \code{"id"}, or
\code{"title"}}

\item{qualifier}{The qualifier variable. In a one-mode network, this
  variable can be used to count only congruence or conflict ties. For
  example, in an organization x organization network via common concepts,
  a binary \code{"agreement"} qualifier could be used to record only ties
  where both organizations have a positive stance on the concept or where
  both organizations have a negative stance on the concept. With an
  integer qualifier, the tie weight between the organizations would be
  proportional to the similarity or distance between the two organizations
  on the scale of the integer variable.

  In a two-mode network, the qualifier variable can be used to retain only
  positive or only negative statements or subtract negative from positive
  mentions. All of this depends on the setting of the
  \code{qualifierAggregation} argument. For event lists, the qualifier
  variable is only used for filtering out duplicates (depending on the
  setting of the \code{duplicate} argument.

  The qualifier can also be \code{NULL}, in which case it is ignored, meaning
  that values in \code{variable1} and \code{variable2} are unconditionally
  associated with each other in the network when they co-occur. This is
  identical to selecting a qualifier variable and setting
  \code{qualifierAggregation = "ignore"}.}

\item{duplicates}{Setting for excluding duplicate statements before network
construction. Valid settings are \code{"include"} (for including all
statements in network construction), \code{"document"} (for counting
only one identical statement per document), \code{"week"} (for counting
only one identical statement per calendar week), \code{"month"} (for
counting only one identical statement per calendar month), \code{"year"}
(for counting only one identical statement per calendar year), and
\code{"acrossrange"} (for counting only one identical statement across
the whole time range).}

\item{start.date}{The start date for network construction in the format
"dd.mm.yyyy". All statements before this date will be excluded.}

\item{stop.date}{The stop date for network construction in the format
"dd.mm.yyyy". All statements after this date will be excluded.}

\item{start.time}{The start time for network construction on the specified
\code{start.date}. All statements before this time on the specified date
will be excluded.}

\item{stop.time}{The stop time for network construction on the specified
\code{stop.date}. All statements after this time on the specified date
will be excluded.}

\item{excludeValues}{A list of named character vectors that contains entries
which should be excluded during network construction. For example,
\code{list(concept = c("A", "B"), organization = c("org A", "org B"))}
would exclude all statements containing concepts "A" or "B" or
organizations "org A" or "org B" when the network is constructed. This
is irrespective of whether these values appear in \code{variable1},
\code{variable2}, or the \code{qualifier}. Note that only variables at
the statement level can be used here. There are separate arguments for
excluding statements nested in documents with certain meta-data.}

\item{excludeAuthors}{A character vector of authors. If a statement is
nested in a document where one of these authors is set in the "Author"
meta-data field, the statement is excluded from network construction.}

\item{excludeSources}{A character vector of sources. If a statement is
nested in a document where one of these sources is set in the "Source"
meta-data field, the statement is excluded from network construction.}

\item{excludeSections}{A character vector of sections. If a statement is
nested in a document where one of these sections is set in the "Section"
meta-data field, the statement is excluded from network construction.}

\item{excludeTypes}{A character vector of types. If a statement is
nested in a document where one of these types is set in the "Type"
meta-data field, the statement is excluded from network construction.}

\item{invertValues}{A boolean value indicating whether the entries provided
by the \code{excludeValues} argument should be excluded from network
construction (\code{invertValues = FALSE}) or if they should be the only
values that should be included during network construction
(\code{invertValues = TRUE}).}

\item{invertAuthors}{A boolean value indicating whether the entries provided
by the \code{excludeAuthors} argument should be excluded from network
construction (\code{invertAuthors = FALSE}) or if they should be the
only values that should be included during network construction
(\code{invertAuthors = TRUE}).}

\item{invertSources}{A boolean value indicating whether the entries provided
by the \code{excludeSources} argument should be excluded from network
construction (\code{invertSources = FALSE}) or if they should be the
only values that should be included during network construction
(\code{invertSources = TRUE}).}

\item{invertSections}{A boolean value indicating whether the entries
provided by the \code{excludeSections} argument should be excluded from
network construction (\code{invertSections = FALSE}) or if they should
be the only values that should be included during network construction
(\code{invertSections = TRUE}).}

\item{invertTypes}{A boolean value indicating whether the entries provided
by the \code{excludeTypes} argument should be excluded from network
construction (\code{invertTypes = FALSE}) or if they should be the
only values that should be included during network construction
(\code{invertTypes = TRUE}).}

\item{method}{This argument represents the clustering method to be used for
the dendrogram. Only hierarchical clustering methods are compatible with
dendrograms. The following values are permitted:
\describe{
  \item{"best"}{Automatically choose the best clustering method with a given
    number of clusters \code{k} (or between 2 and \code{k.max} if
    \code{k = 0}). The selection is based on network modularity of a given
    cluster solution in the subtract network.}
  \item{"single"}{Hierarchical clustering with single linkage.}
  \item{"average"}{Hierarchical clustering with average linkage.}
  \item{"complete"}{Hierarchical clustering with complete linkage.}
  \item{"ward"}{Hierarchical clustering with Ward's algorithm.}
  \item{"fastgreedy"}{Fast & greedy community detection.}
  \item{"walktrap"}{Walktrap community detection.}
  \item{"leading_eigen"}{Leading eigenvector community detection.}
  \item{"edge_betweenness"}{Edge betweenness community detection
    (Girvan-Newman algorithm).}
}}

\item{k}{If \code{method = "best"} is selected, \code{k} is used to determine
at which level the respective clustering method works best. For example, if
\code{k = 3} is supplied, the algorithm will compare the modularity of the
different cluster solutions with three clusters each time to determine the
best-fitting cluster solution. If \code{k = 0} is supplied (the default),
all solutions between one and \code{k.max} clusters will be attempted. The
\code{k} argument also determines the number of rectangles that are drawn
if the \code{rectangle.colors} argument is used. If \code{k = 0}, the
actual number of clusters that works best will be used instead.}

\item{k.max}{The maximal number of clusters to try if \code{method = "best"}
is used.}

\item{rectangle.colors}{If \code{NULL}, no rectangles are drawn. If a single
color is provided (for example, \code{"purple"} or \code{"#AA9900"}),
\code{k} rectangles (one per cluster) will be drawn, and they will all be
in the same color. If \code{k} colors are provided as a vector, each
rectangle will be in a separate color. If \code{rectangle.colors =
"cluster"}, different colors will be picked for the rectangles based on
cluster membership.}

\item{labels}{Which labels to use for variable 1. These can be:
\describe{
  \item{"value"}{The actual variable values for \code{variable1} as extracted
    from the "value" column in a \code{\link{dna_getAttributes}} call.}
  \item{"color"}{The color designated for \code{variable1} in DNA, as
    extracted from the "color" column in a \code{\link{dna_getAttributes}}
    call.}
  \item{"type"}{The type designated for \code{variable1} in DNA, as
    extracted from the "type" column in a \code{\link{dna_getAttributes}}
    call.}
  \item{"alias"}{The alias designated for \code{variable1} in DNA, as
    extracted from the "alias" column in a \code{\link{dna_getAttributes}}
    call.}
  \item{"notes"}{The notes designated for \code{variable1} in DNA, as
    extracted from the "notes" column in a \code{\link{dna_getAttributes}}
    call.}
  \item{a vector of \code{character} objects}{A vector with as many labels
    as there are rows in a \code{\link{dna_getAttributes}} call, in the
    same order (i.e., replacing the original labels that are in
    alphabetical order).}
}}

\item{label.colors}{Colors for the labels. Numbers, string colors, or
hexadecimal strings are allowed. The following values are permitted:
\describe{
  \item{a single color}{A single color for all labels.}
  \item{a vector of \code{k} colors}{A separate color for the labels in
    each cluster.}
  \item{a vector of as many colors as labels}{A separate color for each
    single label.}
  \item{"cluster"}{A separate color is chosen automatically for the labels
    in each cluster. There are \code{k} different colors.}
  \item{"color"}{The colors stored in the "color" column in the attribute
    manager of the DNA database are used, as extracted using the
    \code{\link{dna_getAttributes}} function.}
  \item{"type"}{A separate color for each value in the "type" column in the
    attribute manager of DNA is used.}
  \item{"alias"}{A separate color for each value in the "alias" column in
    the attribute manager of DNA is used.}
  \item{"notes"}{A separate color for each value in the "notes" column in
    the attribute manager of DNA is used.}
}}

\item{label.size}{Font size for the labels.}

\item{label.truncate}{Number of characters to retain for each label. If all
characters should be kept, \code{Inf} can be set.}

\item{leaf.shape}{The way the dendrogram leaves are drawn. The following
values are permitted:
\itemize{
  \item "elbow"
  \item "link"
  \item "diagonal"
  \item "arc"
  \item "fan"
}}

\item{leaf.colors}{The colors of the leaves in the dendrogram. The same
values as in the \code{label.colors} argument are permitted (see the
description above for details).}

\item{leaf.width}{The line width of the leaves.}

\item{leaf.alpha}{The opacity of the leaves and symbols (between 0 and 1).}

\item{symbol.shapes}{The shapes of the leaf end symbols for each leaf. The
following values are permitted:
\describe{
  \item{a single integer}{A single \code{pch} symbol to use for all leaves,
    for example \code{2} or \code{19}.}
  \item{an integer vector with \code{k} elements}{A vector of separate
    \code{pch} symbol integer values for the labels in each cluster. For
    example, if there are three clusters, this could be \code{1:3}.}
  \item{an integer vector with as many elements as symbols}{A vector of
    separate \code{pch} values for each value in \code{variable1}, in the
    order as the rows produced by \code{\link{dna_getAttributes}} (i.e.,
    alphabetical label order).}
  \item{"cluster"}{A separate symbol is automatically selected for each
    cluster.}
  \item{"color"}{A separate symbol is automatically selected for each color
    saved in the "color" column of the output of
    \code{\link{dna_getAttributes}}, which represents the colors in the
    attribute manager in DNA.}
  \item{"type"}{A separate symbol is automatically selected for each type
    saved in the "type" column of the output of
    \code{\link{dna_getAttributes}}, which represents the types in the
    attribute manager in DNA.}
  \item{"alias"}{A separate symbol is automatically selected for each alias
    saved in the "alias" column of the output of
    \code{\link{dna_getAttributes}}, which represents the aliases in the
    attribute manager in DNA.}
  \item{"notes"}{A separate symbol is automatically selected for each note
    saved in the "notes" column of the output of
    \code{\link{dna_getAttributes}}, which represents the notes in the
    attribute manager in DNA.}
}}

\item{symbol.colors}{The colors of the symbols at the leaf ends. The same
values are permitted as in the \code{label.colors} argument (see the
description there).}

\item{symbol.sizes}{The sizes of the symbols at the leaf ends. The default
value is \code{5} for all symbols. Instead of a single number, a vector of
\code{k} values (one for each cluster) or a vector with as many values
as there are leafs or labels can be supplied.}

\item{circular}{Draw a dendrogram with a circular layout?}

\item{theme}{The theme to be used. See \code{\link[ggplot2]{ggtheme}} for
details. Permitted values are:
\itemize{
  \item "bw"
  \item "classic"
  \item "gray"
  \item "dark"
  \item "light"
  \item "minimal"
}}

\item{caption}{Add a caption with details at the bottom of the plot? The
details include the clustering method, the number of clusters, and the
modularity value given the number of clusters with this method.}

\item{return.multiclust}{Instead of returning a \code{ggplot2} plot, return
the \code{dna_multiclust} object upon which the dendrogram is based?}
}
\value{
A \code{ggplot2} plot.
}
\description{
Create a cluster dendrogram based on a DNA database.
}
\details{
This function serves to conduct a cluster analysis of a DNA dataset and
visualize the results as a dendrogram. The user can either select a specific
clustering method or let the \code{\link{dna_multiclust}} function determine
the best cluster solution. The following clustering methods are available for
creating dendrograms:
\itemize{
 \item Hierarchical clustering with single linkage.
 \item Hierarchical clustering with average linkage.
 \item Hierarchical clustering with complete linkage.
 \item Hierarchical clustering with Ward's algorithm.
 \item Fast & greedy community detection.
 \item Walktrap community detection.
 \item Leading eigenvector community detection.
 \item Edge betweenness community detection (Girvan-Newman algorithm).
}
The resulting dendrograms can have different label, leaf, and symbol
properties as well as rectangles for the clusters and other customization
options. It is possible to return the underlying \code{\link{dna_multiclust}}
object instead of the plot by using the \code{return.multiclust} argument.
}
\examples{
library("rDNA")
dna_init()
samp <- dna_sample()
conn <- dna_connection(samp)

# Single-linkage with k = 2 is chosen automatically based on modularity:
dna_dendrogram(conn, method = "best", k = 0)

# Walktrap community detection with three clusters and rectangles:
dna_dendrogram(conn, method = "walktrap", k = 3, rectangle.colors = "purple")

# Custom colors and shapes:
dna_dendrogram(conn,
               label.colors = "color",
               leaf.colors = "cluster",
               rectangle.colors = c("steelblue", "orange"),
               symbol.shapes = 17:18,
               symbol.colors = 3:4)

# Circular dendrogram:
dna_dendrogram(conn, circular = TRUE, label.truncate = 12)

# Modifying the underlying network, e.g., leaving out concepts:
dna_dendrogram(conn, excludeValues = list(concept =
  "There should be legislation to regulate emissions."))

# Return the dna_multiclust object
mc <- dna_dendrogram(conn, k = 0, method = "best", return.multiclust = TRUE)
mc

}
\author{
Philip Leifeld, Johannes B. Gruber
}
