% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_downloadJar}
\alias{dna_downloadJar}
\title{Download the binary DNA JAR file}
\usage{
dna_downloadJar(file = "dna-2.0-beta20.jar", force = FALSE)
}
\arguments{
\item{file}{You can choose to provide a path and name here. By default, the
file is downloaded into the current working directory and the name stays
intact.}

\item{force}{Logical. Should the file be overwritten if it already exists.}
}
\description{
Downloads the newest released DNA JAR file necessary for running
\code{dna_init}.
}
\details{
This simple function downloads the DNA JAR from the latest release.
}
\examples{
dna_downloadJar()
}
