% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_getAttributes}
\alias{dna_getAttributes}
\title{Get the entities and attributes for a variable}
\usage{
dna_getAttributes(
  statementType = NULL,
  variable = NULL,
  statementTypeId = NULL,
  variableId = NULL
)
}
\arguments{
\item{statementType}{The name of the statement type in which the variable is
defined for which entities and values should be retrieved. Only required if
\code{variableId} is not supplied. Either \code{statementType} or
\code{statementTypeId} must be specified in this case.}

\item{variable}{The name of the variable for which the entities and
attributes should be returned. In addition to this argument, either the
statement type name or statement type ID must be supplied to identify the
variable correctly. If the \code{variableId} a specified, the
\code{variable} argument is unnecessary and the statement type need not be
supplied.}

\item{statementTypeId}{The ID of the statement type in which the variable is
defined for which entities and values should be retrieved. Only required if
\code{variableId} is not supplied. Either \code{statementType} or
\code{statementTypeId} must be specified in this case.}

\item{variableId}{The ID of the variable for which the entities and
attributes should be returned. If this argument is supplied, the other
three arguments are unnecessary.}
}
\description{
Retrieve the entities and their attributes for a variable in DNA
}
\details{
This function retrieves the entities and their attributes for a given
variable from the DNA database as a \code{dna_attributes} object. Such an
object is an extension of a data frame and can be treated as such.

There are three ways to use this function: by specifying only the variable
ID; by specifying the variable name and its statement type ID; and by
specifying the variable name and its statement type name.
}
\examples{
\dontrun{
dna_init()
dna_sample()
dna_openDatabase("sample.dna", coderId = 1, coderPassword = "sample")

dna_getAttributes(variableId = 1)
dna_getAttributes(statementTypeId = 1, variable = "organization")
dna_getAttributes(statementType = "DNA Statement", variable = "concept")
}

}
\author{
Philip Leifeld
}
