% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_getCoders}
\alias{dna_getCoders}
\title{Retrieve a dataframe with all coders and their permissions}
\usage{
dna_getCoders(connection)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}
}
\description{
Retrieve a dataframe with all coders and their permissions.
}
\details{
This function creates a dataframe with one row per coder and contains columns
for the ID, name, color, and 12 permissions of the coders. The permissions
serve to govern what coders are allowed to do in the graphical user
interface.
}
