% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_getDocuments}
\alias{dna_getDocuments}
\title{Retrieve documents from the DNA database}
\usage{
dna_getDocuments(documentIds = integer())
}
\arguments{
\item{documentIds}{A vector of document IDs to retrieve. If this argument is
not supplied or is an empty vector, all documents are returned. If this
argument is supplied, only the documents with the given IDs are returned.}
}
\value{
A data frame with the documents of the given IDs. The data frame has
  one row per document and columns for the document ID, coder ID, title,
  text, and date/time, among other variables.
}
\description{
Retrieve documents from the DNA database.
}
\details{
This function retrieves documents from the DNA database and returns them as a
data frame. If no document IDs are specified, all documents are returned. If
document IDs are specified, only those documents are returned. The function
returns a data frame with one row per document and columns for the document
ID, name, text, coder ID, and date/time of creation.
}
\examples{
\dontrun{
dna_init()
dna_sample()
dna_openDatabase(coderId = 1,
                 coderPassword = "sample",
                 db_url = "sample.dna")
documents <- dna_getDocuments()
documents
documents <- dna_getDocuments(documentIds = c(1, 2))
documents
}
}
\seealso{
Other documents: 
\code{\link{dna_addDocuments}()},
\code{\link{dna_deleteDocuments}()},
\code{\link{print.dna_documents}()}
}
\author{
Philip Leifeld
}
\concept{documents}
