% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_getHeadlessDna}
\alias{dna_getHeadlessDna}
\title{Get headless DNA}
\usage{
dna_getHeadlessDna()
}
\value{
Headless DNA reference.
}
\description{
Get headless DNA.
}
\details{
This function returns a reference to the Java class that represents the
R interface of DNA, i.e., the headless API. Headless means that no graphical
user interface is necessary to use the functions in DNA. This is the raw API
interface and should only be used by expert users or to debug other rDNA
functions. You can save the headless DNA object to an object in R and access
its Java functions directly as slots using the "$" operator, for example
\code{dna_getHeadlessDna()$getExport()} will return a reference to the
\code{Exporter} class, which does all the network computations, and
\code{dna_getHeadlessDna()$getAttributes()} will return attributes etc. You
can view the different functions that are available by using the tab key to
auto-complete function access after typing the "$" if the headless reference
has been saved to an object in the workspace.
}
\seealso{
Other {rDNA package startup}: 
\code{\link{dna_init}()},
\code{\link{dna_jar}()},
\code{\link{dna_sample}()}
}
\author{
Philip Leifeld
}
\concept{{rDNA package startup}}
