% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_getRegex}
\alias{dna_getRegex}
\title{Retrieve a dataframe with all regular expressions}
\usage{
dna_getRegex(connection)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}
}
\description{
Retrieve a dataframe with all regular expressions.
}
\details{
The DNA graphical user interface can highlight keywords with colors in the
text. This facilitates the coding process. The \code{dna_getRegex} function
retrieves all regular expressions defined in a DNA database along with their
colors.
}
\author{
Philip Leifeld
}
