% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_getStatements}
\alias{dna_getStatements}
\title{Retrieve statements for a given statement type}
\usage{
dna_getStatements(statementType = 1, statementIds = integer())
}
\arguments{
\item{statementType}{The statement type for which statements should be}

\item{statementIds}{A vector of statement IDs to retrieve. If this argument}
}
\value{
A data frame with the statements of the given type. The data frame
  has one row per statement and columns for the statement ID, document ID,
  start and end positions, coder ID, and the values of the variables defined
  in the statement type.
}
\description{
Retrieve statements for a given statement type.
}
\details{
This function retrieves statements from the DNA database for a given
statement type and returns them as a data frame. The statement type can be
specified by its ID or by its name. If no statement IDs are specified, all
statements of the given type are returned. If statement IDs are specified,
only those statements are returned. The function returns a data frame with
one row per statement and columns for the statement ID, document ID, start
and end positions, coder ID, and the values of the variables defined in the
statement type.
}
\examples{
\dontrun{
dna_init()
dna_sample()
dna_openDatabase(coderId = 1,
                 coderPassword = "sample",
                 db_url = "sample.dna")
statements <- dna_getStatements(statementType = "DNA Statement")
statements
statements <- dna_getStatements(statementType = 1, statementIds = c(1, 2))
statements
}
}
\author{
Philip Leifeld
}
\concept{statements}
