% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_getStatementTypes}
\alias{dna_getStatementTypes}
\title{Retrieve a dataframe with statement types from a DNA connection}
\usage{
dna_getStatementTypes(connection)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}
}
\description{
Retrieve a dataframe with all statement types from a DNA connection.
}
\details{
This function creates a dataframe with one row per statement type and
contains columns for the statement type ID, label, and color (as an RGB hex
string). The statement type IDs can then be used to retrieve the variables
defined within a statement type using the \link{dna_getVariables} function.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())
statementTypes <- dna_getStatementTypes(conn)
}

}
