% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_getVariables}
\alias{dna_getVariables}
\title{Retrieve a dataframe with all variables for a statement type}
\usage{
dna_getVariables(statementType)
}
\arguments{
\item{statementType}{The statement type for which statements should be
retrieved. The statement type can be supplied as an integer or character
string, for example \code{1} or \code{"DNA Statement"}.}
}
\description{
Retrieve a dataframe with all variables defined in a given statement type.
}
\details{
For a given statement type ID or label, this function creates a data frame
with one row per variable and contains columns for the variable ID, name and
data type.
}
\examples{
\dontrun{
dna_init()
samp <- dna_sample()
dna_openDatabase(samp, coderId = 1, coderPassword = "sample")
variables <- dna_getVariables("DNA Statement")
variables
}

}
\author{
Philip Leifeld
}
