% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_jar}
\alias{dna_jar}
\title{Find the DNA jar file}
\usage{
dna_jar()
}
\value{
The file name of the jar file that matches the installed \pkg{rDNA}
  version, including full path.
}
\description{
Find the DNA jar file in the library path or working directory.
}
\details{
rDNA requires the installation of a DNA jar file to run properly. The jar
file is shipped with the rDNA package and is installed in the \code{java/}
directory of the package installation directory in the R library tree. The
version number of the jar file and the rDNA package must match for DNA and
rDNA to be able to work together. The \code{dna_jar} function looks for
the jar file in the package installation directory sub-directory and
returns its file name with its absolute path. If it cannot be found in the
installation directory, the function looks in the current working
directory. The function is also called by \code{\link{dna_init}} if the
location of the jar file is not provided explicitly.
}
\seealso{
Other startup: 
\code{\link{dna_getHeadlessDna}()},
\code{\link{dna_init}()},
\code{\link{dna_sample}()}
}
\author{
Philip Leifeld
}
\concept{startup}
