% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_openConnectionProfile}
\alias{dna_openConnectionProfile}
\title{Open a connection profile}
\usage{
dna_openConnectionProfile(file, coderPassword = "")
}
\arguments{
\item{file}{The file name of the connection profile to open.}

\item{coderPassword}{The clear text coder password. If a zero-length
character object (\code{""}) is provided, the user will be prompted
for a password interactively.}
}
\description{
Open a connection profile and establish a connection to the database.
}
\details{
Load a connection profile from a \code{.dnc} file. The file contains
connection details for a database (like a bookmark) along with the coder ID
of the coder who saved the connection profile. By loading the connection
profile, a connection to the database will be established by DNA, and the
coder saved in the connection profile will be activated. The coder password
the user needs to provide is the coder password for the coder saved in the
connection profile. It serves to decrypt the information stored in the file
and activate the coder in the database connection. If an empty character
object is provided as the password (\code{""}), the user will be prompted
interactively for a password. If the \pkg{askpass} package is installed, this
package will be used to mask the user input; otherwise the password is
visible in clear text. Installing the \pkg{askpass} package is strongly
recommended.
}
\examples{
\dontrun{
dna_init()
dna_sample()
dna_openDatabase(coderId = 1,
                 coderPassword = "sample",
                 db_url = "sample.dna")
dna_saveConnectionProfile(file = "my profile.dnc", coderPassword = "sample")
dna_closeDatabase()
dna_openConnectionProfile(file = "my profile.dnc", coderPassword = "sample")
}

}
\seealso{
Other {rDNA database connections}: 
\code{\link{dna_closeDatabase}()},
\code{\link{dna_openDatabase}()},
\code{\link{dna_printDetails}()},
\code{\link{dna_saveConnectionProfile}()}
}
\author{
Philip Leifeld
}
\concept{{rDNA database connections}}
