% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_openDatabase}
\alias{dna_openDatabase}
\title{Open a database}
\usage{
dna_openDatabase(
  db_url,
  coderId = 1,
  coderPassword = "",
  db_type = "sqlite",
  db_name = "",
  db_port = -1,
  db_login = "",
  db_password = ""
)
}
\arguments{
\item{db_url}{The URL for accessing the database (for remote databases) or
the path of the SQLite database file, including file extension.}

\item{coderId}{The coder ID of the coder who is opening the database. If an
invalid coder ID is supplied (i.e., \code{-1} or similar), the coder ID is
queried interactively from the user.}

\item{coderPassword}{The coder password of the coder who is opening the
database. If an empty password is provided (e.g., \code{""}), the password
is queried interactively from the user.}

\item{db_type}{The type of database. Valid values are \code{"sqlite"},
\code{postgresql}, and \code{postgresql}.}

\item{db_name}{The name of the database at the given URL or path. Can be a
zero-length character object (\code{""}) for file-based SQLite databases.}

\item{db_port}{The connection port for the database connection. No port is
required (\code{db_port = -1}) for SQLite databases. MySQL databases often
use port \code{3306}. PostgreSQL databases often use port \code{5432}. If
\code{db_port = NULL}, one of these default values will be selected based
on the \code{db_type} argument.}

\item{db_login}{The login user name for the database. This is the database
login user name, not the coder name. Can be a zero-length character object
(\code{""}) for SQLite databases.}

\item{db_password}{The password for the database. This is the database
password, not the coder password. Can be a zero-length character object
(\code{""}) for SQLite databases.}
}
\description{
Open a database in DNA.
}
\details{
Open a database in DNA. This can be a SQLite, MySQL, or PostgreSQL database.
The database must already have the table structure required for DNA. You must
provide the coder ID and password along with the database credentials. To
look up coder IDs, use the \code{\link{dna_queryCoders}} function.
}
\examples{
\dontrun{
dna_init()
dna_sample()
dna_openDatabase(coderId = 1,
                 coderPassword = "sample",
                 db_url = "sample.dna")
}

}
\seealso{
\code{\link{dna_queryCoders}}

Other database: 
\code{\link{dna_api}()},
\code{\link{dna_closeDatabase}()},
\code{\link{dna_openConnectionProfile}()},
\code{\link{dna_printDetails}()},
\code{\link{dna_saveConnectionProfile}()}
}
\author{
Philip Leifeld
}
\concept{database}
