% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_plotFrequency}
\alias{dna_plotFrequency}
\title{Plot frequency of statements over time}
\usage{
dna_plotFrequency(connection, of = "agreement", timewindow = "days",
  bar = "stacked", ...)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\link{dna_connection} function.}

\item{of}{A variable which is used to group the bars. Can be "agreement",
"organization", "person", "concept" or NULL to disregard differences.}

\item{timewindow}{Bars represent all statements in a certain timewindow. This
can be "days", "months" or "years".}

\item{bar}{Determines if bars should be stacked ("stacked") or side-by-side
("side").}

\item{...}{Additional arguments passed to \link{dna_network}.}
}
\description{
Plots frequency of statements over time.
}
\details{
This function plots frequency of statements over time from a DNA connection
established with \link{dna_connection}. A second variable can be used to
group the bars in the resulting barplot..
}
\examples{
\dontrun{
dna_downloadJar()
dna_init("dna-2.0-beta21.jar")
conn <- dna_connection(dna_sample())

dna_plotFrequency(connection = conn,
                  of = "agreement",
                  timewindow = "days",
                  bar = "stacked")
}
}
\author{
Johannes B. Gruber
}
