% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_plotMDS}
\alias{dna_plotMDS}
\title{Plots an MDS scatterplot from dna.cluster objects}
\usage{
dna_plotMDS(clust, draw_clusters = character(), alpha = 0.25,
  jitter = c(0.5, 0.7), label = FALSE, label_size = 3.5,
  font_colour = "black", ...)
}
\arguments{
\item{clust}{A \code{dna_cluster} object created by the \link{dna_cluster}
function.}

\item{draw_clusters}{Should clustering be highlighted? Possible values are
"polygon", which draws polygons around clusters derived via the best fit in
\link[cluster]{pam}, "ellipse" uses \link[ggplot2]{stat_ellipse} to draw
ellipses around clusters from \link[cluster]{pam} or "density" to perform a
2D kernel density estimation (see \link[MASS]{kde2d}) and display the
results with contours.}

\item{alpha}{The alpha level of the polygons drawn when draw.clusters =
"polygon".}

\item{jitter}{Takes either one value, to control the width of the jittering
of points, two values to control width and height of the jittering of
points (e.g., c(.l, .2)) or "character()" to turn off the jittering of
points.}

\item{label}{Logical. Should labels be plotted?}

\item{label_size, font_colour}{Control the label size and font colour of the
labels if \code{label = TRUE}, label_size takes numeric values, font_colour
takes a character string with a valid colour value.}

\item{...}{Not used. If you want to add more plot options use \code{+} and
the ggplot2 logic (see example).}
}
\description{
Plots a scatterplot with the results of non-metric multidimensional scaling
performed in \link{dna_cluster}.
}
\details{
This function is a convenience wrapper for using the \code{ggplot2} package
to make a scatterplot of the results of non-metric multidimensional scaling
performed in \link{dna_cluster}. It can also add ellipses of polygons to
highlight clusters.
}
\examples{
\dontrun{
dna_downloadJar()
dna_init("dna-2.0-beta20.jar")
conn <- dna_connection(dna_sample())
clust <- dna_cluster(conn)
mds <- dna_plotMDS(clust,
                   mds = TRUE)
# Flip plot with ggplot2 command
library("ggplot2")
mds +
  coord_flip()
  }
}
\author{
Johannes B. Gruber
}
