% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_plotModularity}
\alias{dna_plotModularity}
\title{Plot modularity values in a \code{dna_multiclust} object}
\usage{
dna_plotModularity(x, only.max = TRUE, nrow = 4, ncol = 4,
  include.y = NULL)
}
\arguments{
\item{x}{A \code{dna_multiclust} object, as created by the
\code{\link{dna_multiclust}} function. Ideally with multiple time points.}

\item{only.max}{Only print the maximal modularity values, as opposed to
facets for the modularity values produced by each separate clustering
method?}

\item{nrow, ncol}{The number of rows and columns for the facet layout if
\code{only.max = FALSE} is set.}

\item{include.y}{Include specific y-axis value in the plot, for example
\code{0}.}
}
\value{
A \pkg{ggplot2} plot.
}
\description{
Plot modularity values in a \code{dna_multiclust} object.
}
\details{
This function serves to plot the maximal modularity values saved in a
\code{dna_multiclust} object (as created by the \code{\link{dna_multiclust}}
function). For example, this can shed light on smoothed bipolarization or
multipolarization over time.

This function replaces the \code{dna_plotTimeWindow} function that was
present in earlier versions of the package.
}
\seealso{
\code{\link{dna_multiclust}}, \code{\link{print.dna_multiclust}}
}
\author{
Philip Leifeld
}
