% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_printDetails}
\alias{dna_printDetails}
\title{Print database details}
\usage{
dna_printDetails()
}
\description{
Print number of documents and statements and active coder.
}
\details{
For the DNA database that is currently open, print the number of documents
and statements, the URL, statement types (and their statement counts), and
the active coder to the console.
}
\examples{
\dontrun{
dna_init()
dna_sample()
dna_openDatabase(coderId = 1,
                 coderPassword = "sample",
                 db_url = "sample.dna")
dna_printDetails()
}

}
\author{
Philip Leifeld
}
