% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_queryCoders}
\alias{dna_queryCoders}
\title{Query the coders in a database}
\usage{
dna_queryCoders(
  db_url,
  db_type = "sqlite",
  db_name = "",
  db_port = NULL,
  db_login = "",
  db_password = ""
)
}
\arguments{
\item{db_url}{The URL or full path of the database.}

\item{db_type}{The type of database. Valid values are \code{"sqlite"},
\code{postgresql}, and \code{postgresql}.}

\item{db_name}{The name of the database at the given URL or path. Can be a
zero-length character object (\code{""}) for file-based SQLite databases.}

\item{db_port}{The connection port for the database connection. No port is
required (\code{db_port = -1}) for SQLite databases. MySQL databases often
use port \code{3306}. PostgreSQL databases often use port \code{5432}. If
\code{db_port = NULL}, one of these default values will be selected based
on the \code{db_type} argument.}

\item{db_login}{The login user name for the database. This is the database
login user name, not the coder name. Can be a zero-length character object
(\code{""}) for SQLite databases.}

\item{db_password}{The password for the database. This is the database
password, not the coder password. Can be a zero-length character object
(\code{""}) for SQLite databases.}
}
\description{
Display the coder IDs, names, and colors present in a DNA database.
}
\details{
Some functions require knowing the coder ID with which changes should be
made. This function queries any database, which does not have to be opened,
for their coder IDs, names, and colors, and returns them as a data frame.
}
\examples{
\dontrun{
dna_init()
dna_sample()
dna_queryCoders("sample.dna")
}

}
\author{
Philip Leifeld
}
