% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_removeCoder}
\alias{dna_removeCoder}
\title{Removes a coder from the database}
\usage{
dna_removeCoder(connection, coder, verbose = TRUE)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\link{dna_connection} function.}

\item{coder}{An integer value denoting the ID of the coder to be removed or
a character object of length 1 indicating the name of the coder to be
removed.}

\item{verbose}{Print details?}
}
\description{
Removes a coder from the database based on its ID or name.
}
\details{
The user provides a connection object and the ID or name of a coder, and the
function will remove the coder from the database. If there are documents or
statements associated with the coder, an error will be printed instead.
}
\author{
Philip Leifeld
}
