% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_removeDocument}
\alias{dna_removeDocument}
\title{Removes a document from the database}
\usage{
dna_removeDocument(connection, id, verbose = TRUE)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\link{dna_connection} function.}

\item{id}{An integer value denoting the ID of the document to be removed. The 
\link{dna_getDocuments} function can be used to look up IDs.}

\item{verbose}{Print details on whether the document could be removed?}
}
\description{
Removes a document from the database based on its ID.
}
\details{
The user provides a connection object and the ID of an existing statement in 
the DNA database, and this statement is removed both from memory and from the 
SQL database.
}
\author{
Philip Leifeld
}
