% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_removeStatementType}
\alias{dna_removeStatementType}
\title{Remove a statement type from the database}
\usage{
dna_removeStatementType(connection, statementType, simulate, verbose)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{statementType}{The statement type to be deleted. The statement type
can be supplied as an integer ID or character string, for example \code{1}
or \code{"DNA Statement"}.}

\item{simulate}{Should the changes only be simulated instead of actually
applied to the DNA connection and the SQL database? This can help to
plan more complex recode operations.}

\item{verbose}{Print details about the recode operations?}
}
\description{
Remove a statement type from the database.
}
\details{
Completely remove a statement type from the database, including all
variables, attributes, and statements that are associated with the statement
type.
}
