% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_removeVariable}
\alias{dna_removeVariable}
\title{Add a new variable to a statement type in the database}
\usage{
dna_removeVariable(connection, statementType = 1, variable,
  simulate = TRUE, verbose = TRUE)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{statementType}{The statement type from which variable should be
deleted. The statement type can be supplied as an integer ID or character
string, for example \code{1} or \code{"DNA Statement"}.}

\item{variable}{The name of the variable as a character object.}

\item{simulate}{Should the changes only be simulated instead of actually
applied to the DNA connection and the SQL database? This can help to
plan more complex recode operations.}

\item{verbose}{Print details about the recode operations?}
}
\description{
Add a new variable to a statement type in the database.
}
\details{
Add a new variable to an existing statement type in the database, based on
the statement type ID or label, a name for the new variable, and a data type
specification.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())
dna_removeVariable(conn, 1, "person")
}

}
