% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_renameStatementType}
\alias{dna_renameStatementType}
\title{Rename a statement type}
\usage{
dna_renameStatementType(connection, statementType, label)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{statementType}{The statement type that should be renamed. The statement
type can be supplied as an integer ID or character string, for example
\code{1} or \code{"DNA Statement"}.}

\item{label}{A descriptive new label for the statement type. For example
\code{"DNA Statement" or "My statement type"}. The label may contain
spaces.}
}
\description{
Rename a statement type by assigning a new label.
}
\details{
This function renames a statement type by replacing its label with a new
label.
}
