% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_sample}
\alias{dna_sample}
\title{Provides a small sample database}
\usage{
dna_sample()
}
\description{
Retrieves the location of a small local .dna sample file.
}
\details{
A small sample database to test the functions of rDNA.
}
\examples{
\dontrun{
dna_init("dna-2.0-beta20.jar")
dna_connection(dna_sample())
}
}
\author{
Johannes Gruber
}
