% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_saveConnectionProfile}
\alias{dna_saveConnectionProfile}
\title{Save a connection profile to a file}
\usage{
dna_saveConnectionProfile(file, coderPassword = "")
}
\arguments{
\item{file}{The file name of the connection profile to save.}

\item{coderPassword}{The clear text coder password. If a zero-length
character object (\code{""}) is provided, the user will be prompted
for a password interactively.}
}
\description{
Save connection profile for the current coder and database to disk
}
\details{
Save the current database URL/path, user name, password, port, database name,
and coder to an encrypted JSON file with the extension \code{.dnc}. This file
is called a connection profile. It serves as a bookmark and saves you from
having to enter and store the full connection details each time you want to
access the database. Please make sure you enter the file name with the
extension. You are asked to provide the coder password of the currently
active coder again, for whom the connection profile is saved. This is just
for security reasons. If you do not provide a coder password (e.g., your
password is a zero-length character object \code{""}), you are asked to enter
the password interactively. If the \pkg{askpass} package is installed, this
package will be used to mask the user input; otherwise the password is
visible in clear text. Installing the \pkg{askpass} package is strongly
recommended.
}
\examples{
\dontrun{
dna_init()
dna_sample()
dna_openDatabase(coderId = 1,
                 coderPassword = "sample",
                 db_url = "sample.dna")
dna_saveConnectionProfile(file = "my profile.dnc", coderPassword = "sample")
}

}
\author{
Philip Leifeld
}
