% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_setDocuments}
\alias{dna_setDocuments}
\title{Recode documents and metadata in the DNA database}
\usage{
dna_setDocuments(connection, documents, removeStatements = FALSE,
  simulate = TRUE, verbose = TRUE)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{documents}{A data frame with the following columns:
\enumerate{
  \item id (integer - can be \code{NA} if a new statement is added),
  \item title (character - the title of the document),
  \item text (character - the document text),
  \item coder (integer - ID of  the coder),
  \item author (character - the document author),
  \item source (character - the document source),
  \item section (character - the document section),
  \item notes (character - the document notes),
  \item type (character - the document type),
  \item date (POSIXct or integer; if integer, the value indicates
  milliseconds since the start of 1970-01-01 - the document date/time).
}
\code{NA} values or \code{-1} values are permitted in the id column. If
these are encountered, a new ID is automatically generated, and the
document is added.}

\item{removeStatements}{If a document is present in the DNA database but not
in the \code{documents} dataframe, the respective document is removed
from the database. However, the document may contain statements. If
\code{removeStatements = TRUE} is set, these statements are removed along
with the respective document. If \code{removeStatements = FALSE} is set,
the statements are not deleted, the document is kept as well, and a
message is printed.}

\item{simulate}{Should the changes only be simulated instead of actually
applied to the DNA connection and the SQL database? This can help to
plan more complex recode operations.}

\item{verbose}{Print details about the recode operations?}
}
\description{
Recode documents and their metadata in a DNA database.
}
\details{
This function takes a dataframe with 10 columns (ID, title, text, coder,
author, sources, section, notes, type, date) as returned by the
\link{dna_getDocuments} function and hands it over to a DNA connection in
order to update the documents in the database based on the contents of the
dataframe. The typical workflow is to retrieve the documents and their
metadata using \link{dna_getDocuments}, manipulating the documents, and then
applying the changes with \link{dna_setDocuments}. Documents that are no
longer in the dataframe are removed from the database; documents in the
dataframe that are not in the database are added to the database; and
contents of existing documents are updated. By default, the changes are only
simulated and not actually written into the database. The user can inspect
the reported changes and then apply the actual changes by setting
\code{simulate = FALSE}.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())
documents <- dna_getDocuments(conn)
documents$title[1] <- "New title for first document"
documents$notes[3] <- "Added a note via rDNA."
documents <- documents[, -5] # removing the fifth document
dna_setDocuments(conn, documents, simulate = TRUE) # apply changes
}

}
\author{
Philip Leifeld
}
