% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_toIgraph}
\alias{dna_toIgraph}
\title{Convert DNA networks to igraph objects}
\usage{
dna_toIgraph(x, weighted = TRUE, attributes = NULL)
}
\arguments{
\item{x}{A dna_network (one- or two-mode).}

\item{weighted}{Logical. Should edge weights be used to create a weighted
graph from the dna_network object.}

\item{attributes}{A data.frame with node/vertex  attributes (for example from
\link{dna_getAttributes}.}
}
\description{
This function can convert objects of class 'dna_network_onemode' or
'dna_network_twomode' to igraph objects.
}
\examples{
# convert without attributes
dna_init()
conn <- dna_connection(dna_sample())
nw <- dna_network(conn, networkType = "onemode")
graph <- dna_toIgraph(nw)


# convert with attributes
# onemode:
nw <- dna_network(conn, networkType = "onemode", variable1 = "organization")
graph <- dna_toIgraph(nw, attributes = dna_getAttributes(conn, variable = "organization"))
plot(graph)

# twomode:
nw2 <- dna_network(conn, networkType = "twomode",
                   variable1 = "organization",
                   variable2 = "concept")
graph2 <- dna_toIgraph(nw2, attributes = rbind(
  dna_getAttributes(conn, variable = "organization"),
  dna_getAttributes(conn, variable = "concept")))

plot(graph2)
}
\author{
Johannes B. Gruber
}
