% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_toNetwork}
\alias{dna_toNetwork}
\title{Convert DNA networks to network objects}
\usage{
dna_toNetwork(x, ...)
}
\arguments{
\item{x}{A dna_network.}

\item{...}{Additional arguments passed to \link[network]{as.network.matrix}.}
}
\description{
This function can convert objects of class 'dna_network_onemode' and
'dna_network_twomode' to network objects as used in the \link{network}
package.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())
nw <- dna_network(conn, networkType = "onemode")
network <- dna_toNetwork(nw)

library("statnet")
plot(network)
nw <- dna_network(conn, networkType = "twomode")
network <- dna_toNetwork(nw)
plot(network,
     displaylabels = TRUE,
     label.cex = 0.5,
     usearrows = FALSE,
     edge.col = "gray")
}
}
