% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_updateCoder}
\alias{dna_updateCoder}
\title{Update the name, color, or permissions of a coder in the DNA database}
\usage{
dna_updateCoder(connection, coder, name = NULL, color = NULL,
  addDocuments = NULL, editDocuments = NULL, deleteDocuments = NULL,
  importDocuments = NULL, viewOthersDocuments = NULL,
  editOthersDocuments = NULL, addStatements = NULL,
  viewOthersStatements = NULL, editOthersStatements = NULL,
  editCoders = NULL, editStatementTypes = NULL, editRegex = NULL,
  verbose = TRUE)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{coder}{The ID (as an integer) or name (as a character object) of the
coder whose values should be updated.}

\item{name}{A character object indicating the name of the new coder. If
\code{NULL}, the current value is retained and no changes are made.}

\item{color}{A character object indicating the color in which the coder
should be displayed in the graphical user interface. The color must be
supplied as a hexadecimal string, for example \code{"#FFFF00"} for yellow.
If \code{NULL}, the current value is retained and no changes are made.}

\item{addDocuments}{Logical: should the coder have the permission to add new
documents via the graphical user interface? If \code{NULL}, the current
value is retained and no changes are made.}

\item{editDocuments}{Logical: should the coder have permission to edit the
meta-data of documents in the graphical user interface? If \code{NULL},
the current value is retained and no changes are made.}

\item{deleteDocuments}{Logical: should the coder have permission to delete
documents from the database in the graphical user interface? If
\code{NULL}, the current value is retained and no changes are made.}

\item{importDocuments}{Logical: should the coder have permission to import
documents into the database via the graphical user interface? If
\code{NULL}, the current value is retained and no changes are made.}

\item{viewOthersDocuments}{Logical: should the coder have permission to view
the documents that were added by other coders? If \code{NULL}, the current
value is retained and no changes are made.}

\item{editOthersDocuments}{Logical: should the coder have permission to edit
the meta-data of documents added by other coders? If \code{NULL}, the
current value is retained and no changes are made.}

\item{addStatements}{Logical: should the coder have permission to add new
statements to the database? If \code{NULL}, the current value is retained
and no changes are made.}

\item{viewOthersStatements}{Logical: should the coder have permission to view
the statements in the graphical user interface that were added by other
coders? If \code{NULL}, the current value is retained and no changes are
made.}

\item{editOthersStatements}{Logical: should the coder have permission to edit
the statements in the graphical user interface that were added by other
coders? If \code{NULL}, the current value is retained and no changes are
made.}

\item{editCoders}{Logical: should the coder have permission to add, remove,
or edit coders in the graphical user interface? If \code{NULL}, the current
value is retained and no changes are made.}

\item{editStatementTypes}{Logical: should the coder have permission to add,
remove, or edit statement types in the graphical user interface? If
\code{NULL}, the current value is retained and no changes are made.}

\item{editRegex}{Logical: should the coder have permission to add, remove, or
edit regular expressions in the graphical user interface? If \code{NULL},
the current value is retained and no changes are made.}

\item{verbose}{Print details?}
}
\description{
Update the name, color, or permissions of a coder in the DNA database.
}
\details{
The \code{dna_updateCoder} function can update the name, color, or any
permissions of an existing coder in a DNA database.
}
\author{
Philip Leifeld
}
