% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_updateSetting}
\alias{dna_updateSetting}
\title{Update a setting related to the GUI}
\usage{
dna_updateSetting(connection, key, value)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{key}{The setting to be updated. Valid keys are
\code{"statementColor"}, \code{"activeCoder"}, \code{"popupWidth"},
\code{"version"}, or \code{"date"}.}

\item{value}{The value to be set. For \code{"statementColor"}, valid values
are \code{"statementType"} or \code{"coder"}. This setting determines by
which criterion statements are colored in the text: either by the color of
the respective statement type or by the color of the active coder. For
\code{"activeCoder"}, valid values are the coder IDs of existing coders
(see \link{dna_getCoders}), but as character objects. This setting ensures
that this coder is activated in the graphical user interface when the
database is opened. For \code{"popupWidth"}, valid values range from
\code{"220"} to \code{"9999"}. This determines the width of the popup
windows in the graphical user interface, in which the variables can be
coded. For \code{"version"}, a version string like \code{"2.0 beta 24"}
is permitted. This indicates the version number that was used to create the
database. \code{"date"} is the date the database was last changed by the
graphical user interface. It is given in the format \code{"YYYY-mm-dd"}.
All values must be supplied as character objects.}
}
\description{
Update a setting related to the graphical user interface.
}
\details{
The \code{dna_updateSetting} function can update a setting related to the
graphical user interface, such as the last active coder, the width of
statement popup windows, or the criterion by which statements are colored in
the text.
}
\author{
Philip Leifeld
}
