% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{.onAttach}
\alias{.onAttach}
\title{Display version number and date when the package is attached}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{The library name.}

\item{pkgname}{The package name.}
}
\description{
Display version number and date when the package is attached.
}
