% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{lvmod}
\alias{lvmod}
\title{Calculate the modularity of a network}
\usage{
lvmod(mat)
}
\arguments{
\item{mat}{A network matrix found e.g. in nw$networks (nw being an object
generated via \link{dna_network}).}
}
\description{
Calculate the modularity of a network retrieved via \link{dna_network}.
}
\details{
Uses the function \link[igraph]{modularity.igraph} to calculate the division
of the network into modules for a network retrieved via \link{dna_network}.
}
\examples{
\dontrun{
dna_downloadJar()
dna_init("dna-2.0-beta20.jar")
conn <- dna_connection(dna_sample())
nw <- dna_network(conn,
                  networkType = "onemode")
modularity <- lvmod(nw)
modularity
}
}
\author{
Philip Leifeld, Johannes B. Gruber
}
