% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{plot.dna_backbone}
\alias{plot.dna_backbone}
\title{Plot diagnostics for \code{dna_backbone} objects}
\usage{
\method{plot}{dna_backbone}(x, ...)
}
\arguments{
\item{x}{A \code{dna_backbone} object.}

\item{...}{Further arguments to be passed along. Currently not in use.}
}
\description{
Plot diagnostics for \code{dna_backbone} objects.
}
\details{
The \link{dna_backbone} function creates \code{dna_backbone} objects by
using a simulated annealing algorithm. This plot method shows visual
diagnostics for the iterations of the algorithm.
}
\author{
Philip Leifeld
}
