% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{plot.dna_scale}
\alias{plot.dna_scale}
\title{Plot a \code{dna_scale} object}
\usage{
\method{plot}{dna_scale}(x, ...)
}
\arguments{
\item{x}{A dna_scale object.}

\item{...}{Further options (currently not used).}
}
\description{
Plot convergence diagnostics of the MCMC chain created by the four
\code{dna_scale} functions.
}
\details{
Plots \code{trace plots} of the \code{dna_scale} MCMC output. For further
convergence diagnostics, see \link{dna_convergenceScale}.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())
dna_scale <- dna_scale1dbin(conn,
                            variable1 = "organization",
                            variable2 = "concept",
                            qualifier = "agreement",
                            threshold = 0.51,
                            mcmc_iterations = 20000,
                            mcmc_burnin = 2000,
                            mcmc_thin = 10,
                            store_variables = "both",
                            drop_min_actors = 1,
                            seed = 12345)
plot(dna_scale)
}

}
\author{
Tim Henrichsen, Johannes B. Gruber
}
