% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_barplot}
\alias{print.dna_barplot}
\title{Print a \code{dna_barplot} object}
\usage{
\method{print}{dna_barplot}(x, trim = 30, attr = TRUE, ...)
}
\arguments{
\item{x}{A \code{dna_barplot} object, as returned by the
\code{\link{dna_barplot}} function.}

\item{trim}{Number of maximum characters to display in entity labels.
Entities with more characters are truncated, and the last character is
replaced by an asterisk (\code{*}).}

\item{attr}{Display attributes, such as the name of the variable and the
levels of the qualifier variable if available.}

\item{...}{Additional arguments. Currently not in use.}
}
\description{
Show details of a \code{dna_barplot} object.
}
\details{
Print the data frame returned by the \code{\link{dna_barplot}} function.
}
\seealso{
Other {rDNA barplots}: 
\code{\link{autoplot.dna_barplot}()},
\code{\link{dna_barplot}()}
}
\author{
Philip Leifeld
}
\concept{{rDNA barplots}}
