% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_cluster}
\alias{print.dna_cluster}
\title{Print the summary of a \code{dna_cluster} object}
\usage{
\method{print}{dna_cluster}(x, ...)
}
\arguments{
\item{x}{A \code{dna_cluster} object.}

\item{...}{Further options (currently not used).}
}
\description{
Show details of a \code{dna_cluster} object.
}
\details{
Print original call to the function, the information from the
\link[stats]{hclust} call and additional variables which can be used for
plotting the object.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample(), verbose = FALSE)
clust.l <- dna_cluster(conn)
clust.l
}
}
