% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_connection}
\alias{print.dna_connection}
\title{Print the summary of a \code{dna_connection} object}
\usage{
\method{print}{dna_connection}(x, ...)
}
\arguments{
\item{x}{A \code{dna_connection} object.}

\item{...}{Further options (currently not used).}
}
\description{
Show details of a \code{dna_connection} object.
}
\details{
Print the number of documents and statements to the console after 
establishing a DNA connection.
}
\examples{
download.file("https://github.com/leifeld/dna/releases/download/v2.0-beta.19/dna-2.0-beta19.jar", 
              destfile = "dna-2.0-beta19.jar", mode = "wb")
download.file("https://github.com/leifeld/dna/releases/download/v2.0-beta.19/sample.dna", 
              destfile = "sample.dna", mode = "wb")
dna_init("dna-2.0-beta19.jar")
conn <- dna_connection("sample.dna", verbose = FALSE)
conn
}
