% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_convergenceScale}
\alias{print.dna_convergenceScale}
\title{Print the summary of a \code{dna_convergenceScale} object}
\usage{
\method{print}{dna_convergenceScale}(x, ...)
}
\arguments{
\item{x}{A \code{dna_convergenceScale} object.}

\item{...}{Further options (currently not used).}
}
\description{
Prints the plots of a \code{dna_convergenceScale} object, when the object 
to be plotted exceeds a meaningful plot on one page.
}
\details{
In case you choose to plot a wide range of actors and/or concepts, which 
exceed a meaningful plot on one page, \code{dna_convergenceScale} uses 
\link[ggforce]{facet_wrap_paginate} under the hood to display the single 
facet plots on several pages.
}
\examples{
\dontrun{
dna_convergenceScale <- dna_convergenceScale(dna_scale,
                                             variable = "both",
                                             method = "trace",
                                             colors = NULL,
                                             nrow = 1,
                                             ncol = 1)
dna_convergenceScale
}
}
\author{
Tim Henrichsen, Johannes B. Gruber
}
