% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_documents}
\alias{print.dna_documents}
\title{Print a \code{dna_documents} object}
\usage{
\method{print}{dna_documents}(x, trim = 10, ...)
}
\arguments{
\item{x}{A \code{dna_documents} object, as returned by the
\code{\link{dna_getDocuments}} function.}

\item{trim}{Number of maximum characters to display per column. Contents with
more characters, such as titles and text, are truncated for more compact
display, and the last character is replaced by an asterisk (\code{*}).}

\item{...}{Additional arguments for the print function.}
}
\description{
Show details of a \code{dna_documents} object, with trimmed column widths.
}
\seealso{
Other documents: 
\code{\link{dna_addDocuments}()},
\code{\link{dna_deleteDocuments}()},
\code{\link{dna_getDocuments}()}
}
\author{
Philip Leifeld
}
\concept{documents}
