% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_multiclust}
\alias{print.dna_multiclust}
\title{Print the summary of a \code{dna_multiclust} object}
\usage{
\method{print}{dna_multiclust}(x, ...)
}
\arguments{
\item{x}{A \code{dna_multiclust} object.}

\item{...}{Further options (currently not used).}
}
\description{
Show details of a \code{dna_multiclust} object.
}
\details{
Print abbreviated contents for the slots of a \code{dna_multiclust} object,
which can be created using the \link{dna_multiclust} function.
}
\seealso{
\code{\link{dna_multiclust}}, \code{\link{dna_plotModularity}}
}
\author{
Philip Leifeld
}
