% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_network_twomode}
\alias{print.dna_network_twomode}
\title{Print a \code{dna_network_twomode} object}
\usage{
\method{print}{dna_network_twomode}(x, trim = 5, attr = TRUE, ...)
}
\arguments{
\item{x}{A \code{dna_network_onemode} object, as returned by the
\code{\link{dna_network}} function.}

\item{trim}{Number of maximum characters to display in row and column labels
of the matrix. Labels with more characters are truncated, and the last
character is replaced by an asterisk (\code{*}).}

\item{attr}{Display attributes, such as the start and stop date and time, the
number of statements on which the matrix is based, the function call and
arguments on which the network matrix is based, and the full labels without
truncation.}

\item{...}{Additional arguments. Currently not in use.}
}
\description{
Show details of a \code{dna_network_twomode} object.
}
\details{
Print a two-mode network matrix and its attributes.
}
\seealso{
Other {rDNA networks}: 
\code{\link{as.matrix.dna_network_onemode}()},
\code{\link{as.matrix.dna_network_twomode}()},
\code{\link{autoplot.dna_network}},
\code{\link{dna_network}()},
\code{\link{dna_tidygraph}()},
\code{\link{print.dna_network_onemode}()}
}
\author{
Philip Leifeld
}
\concept{{rDNA networks}}
