% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_scale1dbin}
\alias{print.dna_scale1dbin}
\title{Print the summary of a \code{dna_scale1dbin} object}
\usage{
\method{print}{dna_scale1dbin}(x, ...)
}
\arguments{
\item{x}{A \code{dna_scale1dbin} object.}

\item{...}{Further options (currently not used).}
}
\description{
Show details of the MCMC chain created by \code{dna_scale1dbin}.
}
\details{
Prints the method that was used for the scaling, the number of actors/concepts 
in the object and their mean.
}
\examples{
\dontrun{
dna_scale <- dna_scale1dbin(connection,
                            variable1 = "organization",
                            variable2 = "concept",
                            qualifier = "agreement",
                            zero_as_na = TRUE,
                            threshold = 0.51,
                            theta_constraints = list(
                            `National Petrochemical & Refiners Association` = "+",
                            `Alliance to Save Energy` = "-"),
                            mcmc_iterations = 20000,
                            mcmc_burnin = 2000,
                            mcmc_thin = 10,
                            mcmc_normalize = TRUE,
                            theta_start = NA,
                            alpha_start = NA,
                            beta_start = NA,
                            theta_prior_mean = 0,
                            theta_prior_variance = 1,
                            alpha_beta_prior_mean = 0,
                            alpha_beta_prior_variance = 0.25,
                            store_variables = "both",
                            drop_constant_items = FALSE,
                            drop_min_actors = 1,
                            verbose = TRUE,
                            seed = 12345)
dna_scale                           
}
}
