% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_scale1dord}
\alias{print.dna_scale1dord}
\title{Print the summary of a \code{dna_scale1dord} object}
\usage{
\method{print}{dna_scale1dord}(x, ...)
}
\arguments{
\item{x}{A \code{dna_scale1dord} object.}

\item{...}{Further options (currently not used).}
}
\description{
Show details of the MCMC chain created by \code{dna_scale1dord}.
}
\details{
Prints the method that was used for the scaling, the number of actors/concepts 
in the object and their mean.
}
\examples{
\dontrun{
dna_scale <- dna_scale1dord(connection,
                            variable1 = "organization",
                            variable2 = "concept",
                            qualifier = "agreement",
                            zero_as_na = TRUE,
                            threshold = 0.6,
                            lambda_constraints = list(
                            `Climate change is caused by greenhouse gases (CO2).` =
                            list(2, "-")),
                            mcmc_iterations = 20000,
                            mcmc_burnin = 2000,
                            mcmc_thin = 10,
                            mcmc_tune = 1.5,
                            mcmc_normalize = FALSE,
                            lambda_start = NA,
                            lambda_prior_mean = 0,
                            lambda_prior_variance = 0.1,
                            store_variables = "organization",
                            drop_constant_items = TRUE,
                            verbose = TRUE,
                            seed = 12345)
dna_scale                           
}
}
