% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_scaleKdbin}
\alias{print.dna_scaleKdbin}
\title{Print the summary of a \code{dna_scaleKdbin} object}
\usage{
\method{print}{dna_scaleKdbin}(x, ...)
}
\arguments{
\item{x}{A \code{dna_scaleKdbin} object.}

\item{...}{Further options (currently not used).}
}
\description{
Show details of the MCMC chain created by \code{dna_scaleKdbin}.
}
\details{
Prints the method that was used for the scaling, the number of actors/concepts 
in the object and their mean.
}
\examples{
\dontrun{
dna_scale <- dna_scaleKdbin(connection,
                            variable1 = "organization",
                            variable2 = "concept",
                            qualifier = "agreement",
                            zero_as_na = TRUE,
                            threshold = 0.6,
                            item_constraints = list(
                            `Climate change is caused by greenhouse gases (CO2).` =
                            list(2, "-"),
                            `Climate change is caused by greenhouse gases (CO2).` =
                            c(3, 0),
                            `CO2 legislation will not hurt the economy.` =
                            list(3, "-")),
                            mcmc_iterations = 20000,
                            mcmc_burnin = 2000,
                            mcmc_thin = 10,
                            alpha_beta_start = NA,
                            alpha_beta_prior_mean = 0,
                            alpha_beta_prior_variance = 1,
                            store_variables = "organization",
                            drop_constant_items = TRUE,
                            verbose = TRUE,
                            seed = 12345)
dna_scale                           
}
}
