% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_statements}
\alias{print.dna_statements}
\title{Print a \code{dna_statements} object}
\usage{
\method{print}{dna_statements}(x, trim = 10, ...)
}
\arguments{
\item{x}{A \code{dna_statements} object, as returned by the
\code{\link{dna_getStatements}} function.}

\item{trim}{Number of maximum characters to display per column. Contents with
more characters, such as organisation names or concepts, are truncated for
more compact display, and the last character is replaced by an asterisk
(\code{*}).}

\item{...}{Additional arguments for the print function.}
}
\description{
Show details of a \code{dna_statements} object, with trimmed column widths.
}
\seealso{
Other statements: 
\code{\link{dna_addStatement}()},
\code{\link{dna_deleteStatements}()},
\code{\link{dna_getStatements}()}
}
\author{
Philip Leifeld
}
\concept{statements}
