% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{trim}
\alias{trim}
\title{Truncate labels}
\usage{
trim(x, n, e = "...")
}
\arguments{
\item{x}{A character string}

\item{n}{Max number of characters to truncate to. Value \code{Inf} turns off
truncation.}

\item{e}{String added at the end of x to signal it was truncated.}
}
\description{
Internal function, used to truncate labels
}
\author{
Johannes B. Gruber
}
