\name{dna.gui}
\alias{dna.gui}
\title{Start DNA GUI}
\description{Start the GUI of the Java software Discourse Network Analyzer.}
\usage{
dna.gui()
}
\details{
The method starts the user interface of the Java software
Discourse Network Analyzer (\acronym{DNA}). All changes in the GUI
have to be saved manually and will not be transferred back
to \R.
}
\seealso{
\code{\link{rDNA}}
\code{\link{dna.init}}
\code{\link{dna.network}}
\code{\link{dna.attributes}}
}
\author{
Philip Leifeld (\url{http://www.philipleifeld.de})
}
\keyword{Programming|programming|interface}
\keyword{Programming|IO}
