% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_addStatementType}
\alias{dna_addStatementType}
\title{Add a new statement type (without variables) to the database}
\usage{
dna_addStatementType(connection, label, color = "#FFFF00", ...)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{label}{A descriptive label for the statement type. For example
\code{"DNA Statement" or "My new statement type"}. The label may contain
spaces.}

\item{color}{A color in the form of a hexadecimal RGB string, such as
\code{"#FFFF00"} for yellow.}

\item{...}{Additional arguments can be added here to define the variables
associated with the statement type. For example,
\code{person = "short text"} or \code{agreement = "boolean"} or multiple
arguments like these separated by comma. The variable names should not
contain any spaces, and the values that indicate the data types should be
of types "short text", "long text", "boolean", or "integer".}
}
\description{
Add a new statement type (without variables) to the database.
}
\details{
Add a new statement type to a database. The statement type contains no
variables but can be populated with variables using the
\link{dna_addVariable} function. Along with the the label used to describe
the statement type, a color needs to be supplied in order to display the
statement type in this color in the GUI (see \link{dna_gui}).
}
