% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_backbone}
\alias{dna_backbone}
\title{Compute and retrieve the backbone and redundant set}
\usage{
dna_backbone(
  penalty = 7.5,
  iterations = 10000,
  statementType = "DNA Statement",
  variable1 = "organization",
  variable1Document = FALSE,
  variable2 = "concept",
  variable2Document = FALSE,
  qualifier = "agreement",
  qualifierDocument = FALSE,
  qualifierAggregation = "subtract",
  normalization = "average",
  duplicates = "document",
  start.date = "01.01.1900",
  stop.date = "31.12.2099",
  start.time = "00:00:00",
  stop.time = "23:59:59",
  excludeValues = list(),
  excludeAuthors = character(),
  excludeSources = character(),
  excludeSections = character(),
  excludeTypes = character(),
  invertValues = FALSE,
  invertAuthors = FALSE,
  invertSources = FALSE,
  invertSections = FALSE,
  invertTypes = FALSE,
  fileFormat = NULL,
  outfile = NULL
)
}
\arguments{
\item{penalty}{The penalty parameter for large backbone sets. The larger the
value, the more strongly larger backbone sets are punished and the smaller
the resulting backbone is. Try out different values to find the right size
of the backbone set. Reasonable values could be \code{5}, \code{7.5}, or
\code{12}, for example.}

\item{iterations}{The number of iterations of the simulated annealing
algorithm. More iterations take more time but may lead to better
optimization results.}

\item{statementType}{The name of the statement type in which the variable
of interest is nested. For example, \code{"DNA Statement"}.}

\item{variable1}{The first variable for network construction. In a one-mode
network, this is the variable for both the rows and columns. In a
two-mode network, this is the variable for the rows only. In an event
list, this variable is only used to check for duplicates (depending on
the setting of the \code{duplicates} argument).}

\item{variable1Document}{A boolean value indicating whether the first
variable is at the document level (i.e., \code{"author"},
\code{"source"}, \code{"section"}, \code{"type"}, \code{"id"}, or
\code{"title"}).}

\item{variable2}{The second variable for network construction. In a one-mode
network, this is the variable over which the ties are created. For
example, if an organization x organization network is created, and ties
in this network indicate co-reference to a concept, then the second
variable is the \code{"concept"}. In a two-mode network, this is the
variable used for the columns of the network matrix. In an event list,
this variable is only used to check for duplicates (depending on the
setting of the \code{duplicates} argument).}

\item{variable2Document}{A boolean value indicating whether the second
variable is at the document level (i.e., \code{"author"},
\code{"source"}, \code{"section"}, \code{"type"}, \code{"id"}, or
\code{"title"}}

\item{qualifier}{The qualifier variable. In a one-mode network, this
  variable can be used to count only congruence or conflict ties. For
  example, in an organization x organization network via common concepts,
  a binary \code{"agreement"} qualifier could be used to record only ties
  where both organizations have a positive stance on the concept or where
  both organizations have a negative stance on the concept. With an
  integer qualifier, the tie weight between the organizations would be
  proportional to the similarity or distance between the two organizations
  on the scale of the integer variable. With a short text variable as a
  qualifier, agreement on common categorical values of the qualifier is
  required, for example a tie is established (or a tie weight increased) if
  two actors both refer to the same value on the second variable AND match on
  the categorical qualifier, for example the type of referral.

  In a two-mode network, the qualifier variable can be used to retain only
  positive or only negative statements or subtract negative from positive
  mentions. All of this depends on the setting of the
  \code{qualifierAggregation} argument. For event lists, the qualifier
  variable is only used for filtering out duplicates (depending on the
  setting of the \code{duplicates} argument.

  The qualifier can also be \code{NULL}, in which case it is ignored, meaning
  that values in \code{variable1} and \code{variable2} are unconditionally
  associated with each other in the network when they co-occur. This is
  identical to selecting a qualifier variable and setting
  \code{qualifierAggregation = "ignore"}.}

\item{qualifierDocument}{A boolean value indicating whether the qualifier
variable is at the document level (i.e., \code{"author"},
\code{"source"}, \code{"section"}, \code{"type"}, \code{"id"}, or
\code{"title"}}

\item{qualifierAggregation}{The aggregation rule for the \code{qualifier}
variable. This must be \code{"ignore"} (for ignoring the qualifier
variable), \code{"congruence"} (for recording a network tie only if both
nodes have the same qualifier value in the binary case or for recording the
similarity between the two nodes on the qualifier variable in the integer
case), \code{"conflict"} (for recording a network tie only if both nodes
have a different qualifier value in the binary case or for recording the
distance between the two nodes on the qualifier variable in the integer
case), or \code{"subtract"} (for subtracting the conflict tie value from
the congruence tie value in each dyad; note that negative values will be
replaced by \code{0} in the backbone calculation).}

\item{normalization}{Normalization of edge weights. Valid settings are
\code{"no"} (for switching off normalization), \code{"average"} (for
average activity normalization), \code{"jaccard"} (for Jaccard coefficient
normalization), and \code{"cosine"} (for cosine similarity normalization).}

\item{duplicates}{Setting for excluding duplicate statements before network
construction. Valid settings are \code{"include"} (for including all
statements in network construction), \code{"document"} (for counting
only one identical statement per document), \code{"week"} (for counting
only one identical statement per calendar week), \code{"month"} (for
counting only one identical statement per calendar month), \code{"year"}
(for counting only one identical statement per calendar year), and
\code{"acrossrange"} (for counting only one identical statement across
the whole time range).}

\item{start.date}{The start date for network construction in the format
\code{"dd.mm.yyyy"}. All statements before this date will be excluded.}

\item{stop.date}{The stop date for network construction in the format
\code{"dd.mm.yyyy"}. All statements after this date will be excluded.}

\item{start.time}{The start time for network construction on the specified
\code{start.date}. All statements before this time on the specified date
will be excluded.}

\item{stop.time}{The stop time for network construction on the specified
\code{stop.date}. All statements after this time on the specified date
will be excluded.}

\item{excludeValues}{A list of named character vectors that contains entries
which should be excluded during network construction. For example,
\code{list(concept = c("A", "B"), organization = c("org A", "org B"))}
would exclude all statements containing concepts "A" or "B" or
organizations "org A" or "org B" when the network is constructed. This
is irrespective of whether these values appear in \code{variable1},
\code{variable2}, or the \code{qualifier}. Note that only variables at
the statement level can be used here. There are separate arguments for
excluding statements nested in documents with certain meta-data.}

\item{excludeAuthors}{A character vector of authors. If a statement is
nested in a document where one of these authors is set in the "Author"
meta-data field, the statement is excluded from network construction.}

\item{excludeSources}{A character vector of sources. If a statement is
nested in a document where one of these sources is set in the "Source"
meta-data field, the statement is excluded from network construction.}

\item{excludeSections}{A character vector of sections. If a statement is
nested in a document where one of these sections is set in the "Section"
meta-data field, the statement is excluded from network construction.}

\item{excludeTypes}{A character vector of types. If a statement is
nested in a document where one of these types is set in the "Type"
meta-data field, the statement is excluded from network construction.}

\item{invertValues}{A boolean value indicating whether the entries provided
by the \code{excludeValues} argument should be excluded from network
construction (\code{invertValues = FALSE}) or if they should be the only
values that should be included during network construction
(\code{invertValues = TRUE}).}

\item{invertAuthors}{A boolean value indicating whether the entries provided
by the \code{excludeAuthors} argument should be excluded from network
construction (\code{invertAuthors = FALSE}) or if they should be the
only values that should be included during network construction
(\code{invertAuthors = TRUE}).}

\item{invertSources}{A boolean value indicating whether the entries provided
by the \code{excludeSources} argument should be excluded from network
construction (\code{invertSources = FALSE}) or if they should be the
only values that should be included during network construction
(\code{invertSources = TRUE}).}

\item{invertSections}{A boolean value indicating whether the entries
provided by the \code{excludeSections} argument should be excluded from
network construction (\code{invertSections = FALSE}) or if they should
be the only values that should be included during network construction
(\code{invertSections = TRUE}).}

\item{invertTypes}{A boolean value indicating whether the entries provided
by the \code{excludeTypes} argument should be excluded from network
construction (\code{invertTypes = FALSE}) or if they should be the
only values that should be included during network construction
(\code{invertTypes = TRUE}).}

\item{fileFormat}{An optional file format specification for saving the
backbone results to a file instead of returning an object. Valid values
are \code{"json"}, \code{"xml"}, and \code{NULL} (for returning the results
instead of writing them to a file).}

\item{outfile}{An optional output file name for saving the resulting
network(s) to a file instead of returning an object.}
}
\description{
Compute and retrieve the backbone and redundant set of a discourse network.
}
\details{
This function applies a simulated annealing algorithm to the discourse
network to partition the set of second-mode entities (e.g., concepts) into a
backbone set and a complementary redundant set.
}
\examples{
\dontrun{
dna_init()
dna_sample()
dna_openDatabase("sample.dna", coderId = 1, coderPassword = "sample")

# compute backbone and redundant set
backbone <- dna_backbone(penalty = 6,
                         iterations = 1000,
                         networkType = "onemode",
                         variable1 = "organization",
                         variable2 = "concept",
                         qualifier = "agreement",
                         qualifierAggregation = "subtract",
                         normalization = "average")

b$backbone # show the set of backbone concepts
b$redundant # show the set of redundant concepts
print(b$backbone_network, attr = FALSE) # show the backbone network
print(b$redundant_network, attr = FALSE) # show the redundant network
print(b$full_network, attr = FALSE) # show the full network

## plot diagnostics (using base R)

# arrange plots in a 2 x 2 grid
par(mfrow = c(2, 2))

# temperature and acceptance probability
plot(x = b$diagnostics$iteration,
     y = b$diagnostics$temperature,
     col = "#a50f15",
     type = "l",
     lwd = 3,
     xlab = "Iteration",
     ylab = "Acceptance probability",
     main = "Temperature and acceptance probability")
lines(x = b$diagnostics$iteration[b$diagnostics$acceptance_prob >= 0],
      y = b$diagnostics$acceptance_prob[b$diagnostics$acceptance_prob >= 0])

# spectral distance between full network and backbone network per iteration
plot(x = b$diagnostics$iteration,
     y = b$diagnostics$penalized_backbone_loss, # wrap in log() if necessary
     type = "l",
     xlab = "Iteration",
     ylab = "Penalized backbone loss",
     main = "Penalized spectral distance between full and reduced network")

# number of concepts in the backbone solution per iteration
plot(x = b$diagnostics$iteration,
     y = b$diagnostics$proposed_backbone_size,
     type = "l",
     col = "blue",
     xlab = "Iteration",
     ylab = "Number of elements",
     main = "Number of elements")
lines(x = b$diagnostics$iteration,
      y = b$diagnostics$current_backbone_size,
      col = "green")
lines(x = b$diagnostics$iteration,
      y = b$diagnostics$optimal_backbone_size,
      col = "red")

# ratio of recent acceptances
plot(x = b$diagnostics$iteration,
     y = b$diagnostics$acceptance_ratio_ma,
     type = "l",
     xlab = "Iteration",
     ylab = "Ratio of acceptances in the last 100 iterations",
     main = "Ratio of acceptances in the last 100 iterations")

## plot diagnostics using ggplot2
require("ggplot2")
require("gridExtra")

# temperature and acceptance probability
g_accept <- ggplot(b$diagnostics, aes(y = temperature, x = iteration)) +
  geom_line(color = "#a50f15") +
  geom_line(data = b$diagnostics[b$diagnostics$acceptance_prob >= 0, ],
            aes(y = acceptance_prob, x = iteration)) +
  ylab("Acceptance probability") +
  xlab("Iteration") +
  ggtitle("Temperature and acceptance probability") +
  theme_bw()

# spectral distance between full network and backbone network per iteration
g_loss <- ggplot(b$diagnostics,
                 aes(y = penalized_backbone_loss, x = iteration)) +
  #coord_trans(y = "log") + # use if necessary
  geom_line() +
  ylab("Penalized backbone loss") +
  xlab("Iteration") +
  ggtitle("Penalized spectral distance between full and reduced network") +
  theme_bw()

# number of concepts in the backbone solution per iteration
d <- data.frame(iteration = rep(b$diagnostics$iteration, 3),
                size = c(b$diagnostics$proposed_backbone_size,
                         b$diagnostics$current_backbone_size,
                         b$diagnostics$optimal_backbone_size),
                Criterion = c(rep("Proposed", nrow(b$diagnostics)),
                              rep("Current iteration", nrow(b$diagnostics)),
                              rep("Best solution", nrow(b$diagnostics))))
g_size <- ggplot(d, aes(y = size, x = iteration, color = Criterion)) +
  geom_line() +
  ylab("Number of elements") +
  xlab("Iteration") +
  ggtitle("Number of elements") +
  theme_bw() +
  theme(legend.position = "bottom")

# ratio of recent acceptances
g_ma <- ggplot(b$diagnostics, aes(y = acceptance_ratio_ma, x = iteration)) +
  geom_line() +
  ylab("Ratio of acceptances in the last 100 iterations") +
  xlab("Iteration") +
  ggtitle("Ratio of acceptances in the last 100 iterations") +
  theme_bw()

# arrange in a 2 x 2 grid and save
x <- grid.arrange(g_accept, g_loss, g_size, g_ma, ncol = 2)
ggsave(x, file = "diagnostics.pdf", width = 15, height = 10)
}

}
\author{
Philip Leifeld, Tim Henrichsen
}
