% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_barplot}
\alias{dna_barplot}
\title{Plot agreement and disagreement}
\usage{
dna_barplot(connection, of = "concept", lab.pos = "Agreement",
  lab.neg = "Disagreement", lab = TRUE, colours = FALSE,
  fontSize = 12, barWidth = 0.6, axisWidth = 1.5, truncate = 40,
  ...)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\link{dna_connection} function.}

\item{of}{Category over which (dis-)agreement will be plotted. Most useful
categories are \code{"concept"} and \code{"actor"} but document categories
can be used.}

\item{lab.pos, lab.neg}{Names for (dis-)agreement labels.}

\item{lab}{Determines whether (dis-)agreement labels and title are displayed.}

\item{colours}{If \code{TRUE}, statement colours will be used to fill the
bars. Not possible for all categories.}

\item{fontSize}{Text size in pts.}

\item{barWidth}{Thickness of the bars. bars will touch when set to \code{1}.
When set to \code{0.5}, space between two bars is the same as thickness of
bars.}

\item{axisWidth}{Thickness of the x-axis which separates agreement from
disagreement.}

\item{truncate}{Sets the number of characters to which axis labels (i.e. the
categories of "of") should be truncated.}

\item{...}{Additional arguments passed to \link{dna_network}.}
}
\description{
Plot agreement and disagreement towards statements.
}
\details{
This function plots agreement and disagreement towards DNA Statements for
different categories such as "concept", "person" or "docTitle". The goal is
to determine the centrality of claims. If, for example, concepts are not very
contested, this may mask the extent of polarization with regard to the other
concepts. It often makes sense to exclude those concept in further analysis.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())

dna_barplot(connection = conn,
            of = "concept",
            colours = FALSE,
            barWidth = 0.5)
}

}
\author{
Johannes B. Gruber
}
