% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_closeDatabase}
\alias{dna_closeDatabase}
\title{Close the open DNA database (if any).}
\usage{
dna_closeDatabase()
}
\description{
Close the DNA database that is currently active (if any).
}
\details{
Close the currently active DNA database and display a message confirming that
the database was closed.
}
\examples{
\dontrun{
dna_init()
dna_sample()
dna_openDatabase(coderId = 1,
                 coderPassword = "sample",
                 db_url = "sample.dna")
dna_closeDatabase()
}

}
\seealso{
Other database: 
\code{\link{dna_api}()},
\code{\link{dna_openConnectionProfile}()},
\code{\link{dna_openDatabase}()},
\code{\link{dna_printDetails}()},
\code{\link{dna_saveConnectionProfile}()}
}
\author{
Philip Leifeld
}
\concept{database}
