% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_cluster}
\alias{dna_cluster}
\title{Cluster network from a DNA connection}
\usage{
dna_cluster(connection, variable = "organization", duplicates = "document",
  clust.method = "ward.D2", attribute1 = "color", attribute2 = "value",
  cutree.k = NULL, cutree.h = NULL, ...)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\link{dna_connection} function.}

\item{variable}{The first variable for network construction  (see
\link{dna_network}). The second one defaults to \code{"concept"} but can be
provided via ... if necessary (see \code{variable2} in
\code{dna_connection}).}

\item{duplicates}{Setting for excluding duplicate statements before network
construction (for details see \link{dna_network}. If exclusion of
duplicates results in a binary matrix, \link[vegan]{vegdist} will be used
instead of \link[stats]{dist} to calculate the dissimilarity matrix.}

\item{clust.method}{The agglomeration method to be used. When set to
\code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"},
\code{"average"}, \code{"mcquitty"}, \code{"median"} or \code{"centroid"}
the respective methods from \link[stats]{hclust} will be used. When set to
\code{"edge_betweenness"}, \code{"leading_eigen"} or \code{"walktrap"}
\link[igraph]{cluster_edge_betweenness},
\link[igraph]{cluster_leading_eigen} or \link[igraph]{cluster_walktrap}
respectively, will be used for clustering.}

\item{attribute1, attribute2}{Which attribute of variable from DNA should be
used to assign colours? There are two sets of colours saved in the
resulting object as \link{dna_plotDendro} has two graphical elements to
distinguish between values: leaf_colours and leaf_ends. Possible values are
\code{"id"}, \code{"value"}, \code{"color"}, \code{"type"}, \code{"alias"}
and \code{"note"}.}

\item{cutree.k, cutree.h}{If cutree.k or cutree.h are provided, the tree from
hierarchical clustering is cut into several groups. See $k$ and $h$ in
\link[stats]{cutree} for details.}

\item{...}{Additional arguments passed to \link{dna_network}. This is
especially useful to set qualifier (defaults to \code{"agreement"}) and
normalization (defaults to \code{"no"}) if non-default values are needed for
clustering. Some other options, like qualifierAggregation are turned off}
}
\description{
Clustering methods for DNA connections.
}
\details{
Perform a cluster analysis based on a DNA connection. Clustering is performed
on a collated two-mode network for cluster methods "ward.D", "ward.D2",
"single", "complete", "average", "mcquitty", "median" and "centroid" or on a
one-mode network with the cluster methods "edge_betweenness", "leading_eigen"
and "walktrap" from the \link{igraph} package. The collated two-mode network
is constructed by retrieving individual networks for each of the qualifiers
levels and combining the results. You can look at this network with
View(clust$network) (with "clust" being the outcome of a call to
dna_cluster()).
}
\examples{
\dontrun{
dna_downloadJar()
dna_init("dna-2.0-beta21.jar")
conn <- dna_connection(dna_sample())

clust.l <- dna_cluster(connection,
                       mds = TRUE)

dna_plotDendro(clust.l)
dna_plotHeatmap(clust.l)
dna_plotMDS(clust.l,
            jitter = c(0.5, 0.7))

}
}
\author{
Johannes B. Gruber
}
