% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_downloadJar}
\alias{dna_downloadJar}
\title{Download the binary DNA JAR file}
\usage{
dna_downloadJar(filename = "dna-2.0-beta20.jar", filepath = character(),
  force = FALSE)
}
\arguments{
\item{filename}{Name of the downloaded Jar.}

\item{filepath}{Download path. Defaults to working directory.}

\item{force}{Logical. Should the file be overwritten if it already exists.}
}
\description{
Downloads the newest released DNA JAR file necessary for running
\code{dna_init}.
}
\details{
This simple function downloads the DNA JAR from the latest release.
}
\examples{
\dontrun{
dna_downloadJar()
}
}
