% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_downloadJar}
\alias{dna_downloadJar}
\title{Download the binary DNA jar file}
\usage{
dna_downloadJar(path = paste0(dirname(system.file(".", package =
  "rDNA")), "/", "extdata"), force = FALSE, returnString = FALSE)
}
\arguments{
\item{path}{Directory path in which the jar file will be stored.}

\item{force}{Logical. Should the file be overwritten if it already exists?}

\item{returnString}{Logical. Return the file name of the downloaded jar file?}
}
\description{
Downloads the newest released DNA jar file necessary for running
\link{dna_init}.
}
\details{
This function uses GitHub's API to download the latest DNA jar file to the
working directory.
}
