% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_getStatements}
\alias{dna_getStatements}
\title{Retrieve a dataframe with statements from a DNA connection}
\usage{
dna_getStatements(connection, statementType)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{statementType}{The statement type for which statements should be
retrieved. The statement type can be supplied as an integer or character
string, for example \code{1} or \code{"DNA Statement"}.}
}
\description{
Retrieve a dataframe with all statements from a DNA connection.
}
\details{
This function creates a dataframe with one row per statement and contains
columns for the statement ID, document ID, start and end position in the
text, statement type ID, coder ID, and all variables. Statements are
retrieved for a specific statement type. The data frame can then be manually
manipulated and returned to the DNA database using the
\link{dna_setStatements} function.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())
statements <- dna_getStatements(conn, statementType = "DNA Statement")
}

}
\author{
Philip Leifeld
}
