% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_gui}
\alias{dna_gui}
\title{Open the DNA GUI}
\usage{
dna_gui(infile = NULL, javapath = NULL, memory = 1024)
}
\arguments{
\item{infile}{The file name of the .dna database or the URL of the mySQL 
database to load upon start-up of the GUI}

\item{javapath}{The path to the \code{java} command. This may be useful if 
the CLASSPATH is not set and the java command can not be found. Java 
is necessary to start the DNA GUI.}

\item{memory}{The amount of memory in megabytes to allocate to DNA, for 
example \code{1024} or \code{4096}.}
}
\description{
Start DNA and optionally load a database.
}
\details{
Start the DNA GUI. Optionally load a .dna database or a mySQL online 
database upon start-up of the GUI. This function is useful to use 
DNA on the fly to quickly recode statements or look something up.
}
\examples{
download.file("https://github.com/leifeld/dna/releases/download/v2.0-beta.19/dna-2.0-beta19.jar", 
              destfile = "dna-2.0-beta19.jar", mode = "wb")
dna_init("dna-2.0-beta19.jar")
dna_gui()
}

