% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_init}
\alias{dna_init}
\title{Initialize the connection with DNA}
\usage{
dna_init(jarfile = "dna-2.0-beta20.jar", memory = 1024)
}
\arguments{
\item{jarfile}{The file name of the DNA jar file, e.g.,
\code{"dna-2.0-beta20.jar"}.}

\item{memory}{The amount of memory in megabytes to allocate to DNA, for
example \code{1024} or \code{4096}.}
}
\description{
Establish a connection between \pkg{rDNA} and the DNA software.
}
\details{
To use \pkg{rDNA}, DNA first needs to be initialized. This means that
\pkg{rDNA} needs to be told where the DNA executable file, i.e., the JAR
file, is located. When the \code{dna_init} function is used, the connection
to the DNA software is established, and this connection is valid for the
rest of the \R session. To initialize a connection with a different DNA
version or path, the \R session would need to be restarted first.
}
\examples{
\dontrun{
dna_init("dna-2.0-beta20.jar")
}
}
