% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_plotCluster}
\alias{dna_plotCluster}
\title{Plots cluster objects}
\usage{
dna_plotCluster(dend, shape = "elbows", colour = TRUE, activity = TRUE,
  lineends = TRUE, colours = "identity", custom.colours = character(),
  branch.colour = "#636363", leaf.linetype = "a", branch.linetype = "a",
  lineWidth = 1, lineAlpha = 1, endsAlpha = 1, fontSize = 12,
  theme = "bw", truncate = 40, leaf.labels = "ticks", circular = FALSE,
  show.legend = FALSE, ...)
}
\arguments{
\item{dend}{A \code{dna_cluster} object created by the
\link{dna_cluster} function.}

\item{shape}{See \link[ggraph]{layout_dendrogram_auto}.}

\item{colour}{Should the plot be coloured or not (logical).}

\item{activity}{Should activity of variable in \link{dna_cluster} be used to determine size of lineends (logical).}

\item{lineends}{Should lineends be displayed (logical).}

\item{colours}{There are three options from where to derive the colours in
the plot: (1.) "identity" uses the names of variables as colours, fails if
names are not plottable colours; (2.) "manual" provide colours via
custom.colours; (3.) "brewer" automatically select nice colours from a
\code{RColorBrewer} palette (palette can be set in custom.colours).}

\item{custom.colours}{Either provide enough colours to manually set the
colours in the plot (if colours = "manual") or select a palette from
\code{RColorBrewer} (if colours = "brewer").}

\item{branch.colour}{Provide one colour in which all branches are coloured.}

\item{leaf.linetype, branch.linetype}{"a" for straight line or "b" for dotted line.}

\item{lineWidth}{Width of all lines.}

\item{lineAlpha}{Alpha of all lines.}

\item{endsAlpha}{Alpha of all lineends}

\item{fontSize}{Set the font size for the entire plot.}

\item{theme}{See themes in \code{ggplot2}. The theme "bw" was customised to look best with dendrograms.}

\item{truncate}{Sets the number of characters to which labels should be truncated.}

\item{leaf.labels}{Either "ticks" to display the labels as axis ticks or "node" to label nodes directly.}

\item{circular}{Logical. Should the layout be transformed to a circular representation. See  \link[ggraph]{layout_dendrogram_auto}.}

\item{show.legend}{Logical. Should the legend be displayed.}

\item{...}{additional arguments passed to \link{dna_network}}
}
\description{
Plots objects derived via \link{dna_cluster}.
}
\details{
This function is a convenience wrapper for several different dendrogram
types, which can be plotted using the \code{ggraph} package.
}
\examples{
\dontrun{
dna_init("dna-2.0-beta20.jar")
conn <- dna_connection(dna_sample())
at <- dna_attributes(conn, "DNA Statement", "organization")
dna_plotCentrality(conn)
}
}
\author{
Johannes B. Gruber
}
