% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_plotFrequency}
\alias{dna_plotFrequency}
\title{Plots frequency of statements over time.}
\usage{
dna_plotFrequency(connection, of = "agreement", timewindow = "days",
  bar = "stacked", ...)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\link{dna_connection} function.}

\item{of}{A variable which is used to group the bars. Can be
\code{"agreement"}, \code{"organization"}, \code{"person"},
\code{"concept"} or \code{NULL} to disregard differences.}

\item{timewindow}{Bars represent all statements made during certain
timewindow. This can be \code{"days"}, \code{"months"} or \code{"years"}.}

\item{bar}{Determines if bars should be stacked (\code{"stacked"}) or
side-by-side (\code{"side"}).}

\item{...}{Additional arguments passed to \link{dna_network}.}
}
\description{
This function plots frequency of statements over time from a DNA connection
established with \link{dna_connection}. A second variable can be used to
group the bars in the resulting barplot..
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())

dna_plotFrequency(connection = conn,
                  of = "agreement",
                  timewindow = "days",
                  bar = "stacked")
}
}
\author{
Johannes B. Gruber
}
