% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_plotHeatmap}
\alias{dna_plotHeatmap}
\title{Plots a heatmap from dna_cluster objects}
\usage{
dna_plotHeatmap(clust, truncate = 40, values = FALSE,
  colours = character(), custom_colours = character(), square = TRUE,
  dendro_x = TRUE, dendro_x_size = 0.2, dendro_y_size = 0.2,
  qualifierLevels = list(`0` = "no", `1` = "yes"), show_legend = TRUE,
  ...)
}
\arguments{
\item{clust}{A \code{dna_cluster} object created by the \link{dna_cluster}
function.}

\item{truncate}{Sets the number of characters to which labels should be
truncated. Value \code{Inf} turns off truncation.}

\item{values}{If TRUE, will display the values in the tiles of the heatmap.}

\item{colours}{There are two options: When \code{"brewer"} is selected, the function
\link[ggplot2]{scale_fill_distiller} is used to colour the heatmap tiles.
When \code{"gradient"} is selected, \link[ggplot2]{scale_fill_gradient} will be
used. The colour palette and low/high values can be supplied using the
argument \code{custom_colours}.}

\item{custom_colours}{For \code{colours = "brewer"} you can use either a
string with a palette name or the index number of a brewer palette (see
details). If \code{colours = "gradient"} you need to supply at least two
values. Colours are then derived from a sequential colour gradient palette.
\link[ggplot2]{scale_fill_gradient}. If more than two colours are provided
\link[ggplot2]{scale_fill_gradientn} is used instead.}

\item{square}{If TRUE, will make the tiles of the heatmap quadratic.}

\item{dendro_x}{If TRUE, will draw a dendrogram on the x-axis.}

\item{dendro_x_size, dendro_y_size}{Control the size of the dendrograms on the
x- and y-axis.}

\item{qualifierLevels}{Takes a list with integer values of the qualifier
levels (as characters) as names and character values as labels (See
example).}

\item{show_legend}{Logical. Should a legend be displayed.}

\item{...}{Additional arguments passed to \link{dna_plotDendro}.}
}
\description{
Plots a heatmap with dendrograms from objects derived via \link{dna_cluster}.
}
\details{
This function plots a heatmap including dendrograms on the x- and y-axis of
the heatmap plot. The available options for colouring the tiles can be
displayed using \code{RColorBrewer::display.brewer.all()} (RColorBrewer needs
to be installed).
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())
clust <- dna_cluster(conn)
dend <- dna_plotHeatmap(clust,
qualifierLevels = list("0" = "no", "1" = "yes"))
}
}
\author{
Johannes B. Gruber
}
