% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_plotNetwork}
\alias{dna_plotNetwork}
\title{Plots a network from DNA data}
\usage{
dna_plotNetwork(x, layout = "auto", edges = "link", edge_weight = TRUE,
  edge_size_range = c(0.2, 2), edge_colour = "grey", edge_alpha = 1,
  node_size = 6, node_attribute = "type", node_colours = "auto",
  custom_colours = character(), node_shape = 19, node_label = TRUE,
  font_size = 6, theme = "graph", label_repel = 0.5,
  label_lines = FALSE, truncate = 30, groups = list(), threshold = NULL,
  seed = 12345, show_legend = TRUE, ...)
}
\arguments{
\item{x}{Either a dna_network or a dna_cluster object or a connection to a
DNA database (see \link{dna_connection}).}

\item{layout}{The type of layout to use.}

\item{edges}{When set to "link", the default, straight lines are used to
connect nodes. Other available options are "arc", "diagonal" and "fan".}

\item{edge_weight}{If TRUE, edge weights will be used to determine width of
the lines between nodes. The minimum and maximum width can be controlled
with \code{edge_size_range}.}

\item{edge_size_range}{Takes a numeric vector with two values: minimum and
maximum \code{edge_weight}.}

\item{edge_colour}{What colour to use for edges.}

\item{edge_alpha}{Values lower than 1 make the edges transparent.}

\item{node_size}{Takes numeric values to control the size of nodes.}

\item{node_attribute}{Takes the name of an attribute in DNA or "group" to
colour nodes. The option "group" only makes sense if either a dna_cluster
object is used or if a list of groups is provided in the \code{groups}
argument.}

\item{node_colours}{There are five options for colouring the nodes: (1.)
"auto" simply leaves the standard ggplot2 colours; (2.) "identity" tries to
use the names of variables as colours (i.e., if you set axis = "color") but
fails if names are not plottable colours; (3) "manual" let's you provide
colours via custom_colours; (4.) "brewer" automatically selects nice
colours from a \code{RColorBrewer} palette (palettes can be set in
custom_colours, defaults to \code{Set3}); and (5.) "single" uses the first
value in custom_colours for all nodes.}

\item{custom_colours}{Either provide enough colours to manually set the
colours in the plot (if colours = "manual"), select a palette from
\code{RColorBrewer} (if colours = "brewer") or provide a single colour for
all nodes (if colours = "single").}

\item{node_shape}{Control the node shape. Available shapes range from 0:25
and 32:127.}

\item{node_label}{If TRUE, text is added next to nodes to label them. If
"label", a rectangle is drawn underneath the text, often making it easier
to read.}

\item{font_size}{Control the font size of the node labels.}

\item{theme}{See themes in \code{ggplot2}. Available options are "graph"
(which is customised to look best with networks), "bw", "void", "light" and
"dark". Leave empty to use standard ggplot theme. Customise the theme by
adding \code{+ theme_*} after this function.}

\item{label_repel}{Controls how far from the labels will be put from nodes.
The exact position of nodes is random but overplotting is avoided.}

\item{label_lines}{If TRUE, draws lines between nodes and labels if labels
are further away from nodes.}

\item{truncate}{Sets the number of characters to which labels should be
truncated.}

\item{groups}{Takes a named list with values as names and group assignments
as values.}

\item{threshold}{Minimum threshold for which edges should be plotted.}

\item{seed}{Set seed to reproduce the layout.}

\item{show_legend}{Logical. Should a legend be displayed.}

\item{...}{Only active when plotting from a dna_connection. Arguments are
passed to \link{dna_network}. Especially useful for normalisation of
networks.}
}
\description{
Plots a network from various objects containing DNA data
}
\details{
This function is a convenience wrapper to plot networks  with
\link[ggraph]{ggraph} from several types of objects created in rDNA.
Specifically, one- and twomode networks from \link{dna_network}, dna_cluster
objects created with \link{dna_cluster} and connections to a database
(\link{dna_connection}) are supported. The \code{...} are inactive, except
when plotting directly from a dna_connection. In this case, additional
arguments can be passed to network construction via \link{dna_network}.

The available layouts are listed in \link[ggraph]{layout_igraph_auto} under
"Standard layouts".
}
\examples{
\dontrun{
dna_downloadJar()
dna_init()
conn <- dna_connection(dna_sample())

# Plot from onemode network (twomode also possible)
nw <- dna_network(conn, networkType = "onemode")
dna_plotNetwork(nw)

# Plot from dna_cluster
clust <- dna_cluster(conn)
dna_plotNetwork(clust, node_attribute = "group")

# Plot from dna_connection
dna_plotNetwork(conn, networkType = "twomode", layout = "bipartite")
}

}
