% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_plotPolarization}
\alias{dna_plotPolarization}
\title{Plot polarization values in a \code{dna_polarization} object}
\usage{
dna_plotPolarization(x, convergence = FALSE)
}
\arguments{
\item{x}{A \code{dna_polarization} object, as created by the
\code{\link{dna_polarization}} function. Must have multiple time points for
visualizing the results.}

\item{convergence}{Plot convergence diagnostics for the genetic algorithm.}
}
\value{
A \pkg{ggplot2} plot.
}
\description{
Plot polarization values in a \code{dna_polarization} object.
}
\details{
This function serves to plot the polarization values saved in a
\code{dna_polarization} object (as created by the
\code{\link{dna_polarization}} function). For example, this can shed light on
smoothed bipolarization or multipolarization over time. Note that this only
works when using the time window arguments in the
\code{\link{dna_polarization}} function.
}
\seealso{
\code{\link{dna_polarization}}
}
\author{
Philip Leifeld
}
