% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_plotScale}
\alias{dna_plotScale}
\title{Plot ideological ideal points from a dna_scale object}
\usage{
dna_plotScale(dna_scale, variable = "organization", dimensions = 1,
  hpd = 0.95, colors = TRUE, label = TRUE, label_size = 10,
  point_size = 4, hpd_size = 0.8, intercept = TRUE, facet = NULL,
  truncate = 40, x_axis_label = NULL, y_axis_label = NULL, min = 1,
  exclude = NULL)
}
\arguments{
\item{dna_scale}{A \code{dna_scale} object.}

\item{variable}{Variable to be plotted.}

\item{dimensions}{Number of dimensions to be plotted. To plot only the first or second
dimension of a two dimensional scaling please use \code{2.1} or \code{2.2}. Only two
dimensions are supported.}

\item{hpd}{A numeric scalar specifying the size of the Highest Posterior Density intervals.
Defaults to \code{0.95}.}

\item{colors}{Logical. Should the colors stored in the \code{dna_attributes} be used
for plotting? Defaults to \code{TRUE}.}

\item{label}{Logical. Should labels be plotted? Defaults to \code{TRUE}.}

\item{label_size}{Label size in pts.}

\item{point_size}{Point size in pts.}

\item{hpd_size}{Highest Posterior Density intervals size in pts.}

\item{intercept}{Logical. Should intercept be plotted?}

\item{facet}{A character vector specifying which attribute of the actors/concepts should
be used for the facet plot.}

\item{truncate}{Sets the number of characters to which labels should be truncated.}

\item{x_axis_label}{A character vector specifying the x axis label name.}

\item{y_axis_label}{A character vector specifying the y axis label name.}

\item{min}{A numeric value reducing the plot to actors/concepts with a minimum frequency
of statements.}

\item{exclude}{A character vector to exclude actors/concepts from the plot.}
}
\description{
Plots ideological ideal points of actors and concepts with the results of the MCMC scaling
performed in the \code{dna_scale} functions.
}
\details{
This function is a convenience wrapper for the \code{ggplot2} package to plot ideological
ideal points from \code{dna_scale} objects. Only two dimensions can currently be plotted.
Highest Posterior Densities can be added to the ideal points.

Plotting all actors or concepts can create chaotic plots, you can therefore limit the plot
by including only the most active actors or most mentioned concepts with \code{min}.
Specific entries can be excluded with the \code{exclude} argument. Furthermore, the plot
can be split into several facets according to the attributes of the actors and/or concepts.
}
\examples{
\dontrun{
dna_plotScale(dna_scale,
              variable = "organization",
              dimensions = 1,
              hpd = 0.95,
              label = TRUE,
              label_size = 10,
              point_size = 4,
              hpd_size = 0.8,
              intercept = TRUE,
              facet = "type",
              truncate = NULL,
              x_axis_label = "Ideological Ideal Point",
              min = 1,
              exclude = c("Environmental Protection Agency", "Senate"))

dna_plotScale(dna_scale,
              variable = "concept",
              dimensions = 2.1,
              hpd = 0.95,
              label = TRUE,
              label_size = 10,
              point_size = 4,
              hpd_size = 0.8,
              intercept = TRUE,
              facet = NULL,
              truncate = 40,
              x_axis_label = "Ideological Ideal Point",
              min = 1,
              exclude = "There should be legislation to regulate emissions.")
              }

}
\author{
Tim Henrichsen, Johannes B. Gruber
}
