% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_plotTimeWindow}
\alias{dna_plotTimeWindow}
\title{Plot \link{dna_timeWindow} objects}
\usage{
dna_plotTimeWindow(x, facetValues = "all", include.y = NULL,
  rows = NULL, cols = NULL, diagnostics = FALSE, ...)
}
\arguments{
\item{x}{A \code{dna_timeWindow} object created by the \link{dna_timeWindow}
function.}

\item{facetValues}{The name or names of the facet values which should be
included in the plot.}

\item{include.y}{Include specific value of facet in the plot.}

\item{rows, cols}{Number of rows and columns in which the plots are arranged.
plot.}

\item{diagnostics}{Plot the different measures saved in the object?}

\item{...}{Currently not used. Additional parameters should be passed on to ggplot2 via
e.g. \code{+ theme_bw()}.}
}
\description{
Plot \link{dna_timeWindow} objects in a grid separated by facet.
}
\details{
A convenience function to plot an object created with \link{dna_timeWindow}
function. Uses \link[ggplot2]{geom_line} under the hood to plot results from
a call to \link{dna_timeWindow} and facets a grid view using
\link[ggplot2]{facet_grid}. Customised themes and ggplot2 functions can be
passed on with \code{+}.
}
\examples{
\dontrun{
library("ggplot2")
dna_init()
conn <- dna_connection(dna_sample())

tW <- dna_timeWindow(connection = conn,
                     timewindow = "days",
                     windowsize = 10,
                     facet = "Authors",
                     facetValues = c("Bluestein, Joel",
                                     "Voinovich, George"),
                     method = "bipolarization",
                     verbose = TRUE)

dna_plotTimeWindow(tW,
                   facetValues = c("Bluestein, Joel",
                                   "Voinovich, George",
                                   "all"),
                   rows = 2)

mp <- dna_timeWindow(connection = conn,
                     timewindow = "days",
                     windowsize = 15,
                     method = "multipolarization",
                     duplicates = "document",
                     parallel = "PSOCK",
                     ncpus = 3)

dna_plotTimeWindow(mp, include.y = c(-1, 1)) + theme_bw()
}
}
\author{
Johannes B. Gruber, Philip Leifeld
}
