% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_renameVariable}
\alias{dna_renameVariable}
\title{Rename a variable}
\usage{
dna_renameVariable(connection, statementType, variable, label,
  simulate = TRUE, verbose = TRUE)
}
\arguments{
\item{connection}{A \code{dna_connection} object created by the
\code{dna_connection} function.}

\item{statementType}{The statement type in which the variable is defined that
should be renamed. The statement type can be supplied as an integer ID or
character string, for example \code{1} or \code{"DNA Statement"}.}

\item{variable}{The name of the variable that should be renamed.}

\item{label}{A descriptive new label for the variable. For example
\code{"actor" or "intensity"}. The label must not contain spaces.}

\item{simulate}{Should the changes only be simulated instead of actually
applied to the DNA connection and the SQL database? This can help to
plan more complex recode operations.}

\item{verbose}{Print details about the recode operations?}
}
\description{
Rename a variable by assigning a new label.
}
\details{
This function renames a statement type by replacing its label with a new
label.
}
