% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_sample}
\alias{dna_sample}
\title{Provides a small sample database}
\usage{
dna_sample(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Logical. Should \code{sample.dna} be overwritten if found in
the current working directory?}
}
\description{
A small sample database to test the functions of rDNA.
}
\details{
Copies a small .dna sample file to the current working directory and returns
the location of this newly created file.
}
\examples{
\dontrun{
dna_init()
s <- dna_sample()
dna_openDatabase(s)
}

}
\author{
Johannes B. Gruber, Philip Leifeld
}
