% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_sample}
\alias{dna_sample}
\title{Provides a small sample database}
\usage{
dna_sample(overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{overwrite}{Logical. Should sample.dna be overwritten if found in the
current working directory?}

\item{verbose}{Display warning message if file exists in current wd.}
}
\description{
Copies a small .dna sample file to the current working directory and returns
the location of this newly created file.
}
\details{
A small sample database to test the functions of rDNA.
}
\examples{
\dontrun{
dna_init()
dna_connection(dna_sample())
}

}
\author{
Johannes B. Gruber
}
