% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_setAttributes}
\alias{dna_setAttributes}
\title{Set entities and attributes for a variable}
\usage{
dna_setAttributes(
  data,
  statementType = NULL,
  variable = NULL,
  statementTypeId = NULL,
  variableId = NULL,
  simulate = TRUE
)
}
\arguments{
\item{data}{A data frame of class \code{dna_attributes} or compatible
\code{data.frame}.}

\item{statementType}{The name of the statement type in which the variable is
defined. Only required if \code{variableId} and \code{statementTypeId} are
not supplied.}

\item{variable}{The name of the variable whose entities and attributes should
be updated. Required unless \code{variableId} is supplied.}

\item{statementTypeId}{The ID of the statement type in which the variable is
defined. Only required if \code{variableId} and \code{statementType} are
not supplied.}

\item{variableId}{The ID of the variable whose entities and attributes should
be updated. Required unless \code{variable} and either \code{statementType}
or \code{statementTypeId} is supplied.}

\item{simulate}{Logical; if \code{TRUE}, all changes are rolled back after
execution. Use this as a precaution to test how changes would affect the
database before repeating the function without simulation.}
}
\description{
Update entities (values, colors, and attribute values for a variable in DNA
by supplying a data frame. The data frame must contain the same structure as
returned by \code{\link{dna_getAttributes}}.
}
\details{
The first three columns must be:
\enumerate{
  \item \code{ID} – integer entity ID
  \item \code{value} – character entity value
  \item \code{color} – character hex RGB color (e.g. \code{"#AABBCC"})
}
All remaining columns are interpreted as attribute variables.

Identify which variable to update by supplying one of the following:
\enumerate{
  \item the variable ID
  \item the variable name as a character object and the statement type ID
  \item the variable name and statement type as character objects
}
}
\seealso{
Other attributes: 
\code{\link{dna_getAttributes}()}
}
\author{
Philip Leifeld
}
\concept{attributes}
