% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_timeWindow}
\alias{dna_timeWindow}
\title{Computes for a temporal sequence of networks}
\usage{
dna_timeWindow(connection, timewindow = "days", windowsize = 100,
  facet = "Types", facetValues = character(), method = "modularity",
  verbose = 2, ...)
}
\arguments{
\item{connection}{A \link{dna_connection} object created by the
\link{dna_connection} function.}

\item{timewindow}{Same as in \link{dna_network}.}

\item{windowsize}{Same as in \link{dna_network}.}

\item{facet}{Which value from the dna database should be used to subset the
networks. Can be "Authors" for document author, "Sources" for document
source, "Sections" for documents which contain a certain section or "Types"
to subset document types.}

\item{facetValues}{Which values should be used to facet calculation of the
networks. Always contains the value 'all' for comparison. Use e.g.
excludeTypes to exclude documents from comparison.}

\item{method}{Is used to compute exactly one measurement for each network
computed in the temporal sequence of networks. Can contain the name of any
function which reduces a matrix to just one value.}

\item{verbose}{Display messages if TRUE or 1. Also display messages details
of network construction when 2}

\item{...}{Additional arguments passed to \link{dna_network}.}
}
\description{
Computes a measure for each network in a temporal sequence of networks.
}
\details{
This function serves as a convenience wrapper to calculate a measure for each
network in a temporal sequence of networks. The standard is to calculate the
modularity of the network for each time window (see
\link[igraph]{modularity.igraph}). The function can also be used to split
your data (facet) and calculate networks for each facet type.
}
\examples{
\dontrun{
dna_downloadJar()
dna_init("dna-2.0-beta21.jar")
conn <- dna_connection(dna_sample())

tW <- dna_timeWindow(connection = conn,
                     timewindow = "days",
                     windowsize = 10,
                     facet = "Authors",
                     facetValues = c("Bluestein, Joel",
                                     "Voinovich, George",
                                     "Whitman, Christine Todd"),
                     method = "modularity",
                     verbose = TRUE)

dna_plotTimeWindow(tW, facetValues = c("Bluestein, Joel", "Voinovich, George"))
}
}
\author{
Johannes B. Gruber, Philip Leifeld
}
