% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_timeWindow}
\alias{dna_timeWindow}
\title{Computes a temporal sequence of networks}
\usage{
dna_timeWindow(connection, method = "bipolarization",
  timewindow = "days", windowsize = 100, facet = "Types",
  facetValues = character(), parallel = c("no", "FORK", "PSOCK"),
  ncpus = 1, cl = NULL, verbose = 1, ...)
}
\arguments{
\item{connection}{A \link{dna_connection} object created by the
\link{dna_connection} function.}

\item{method}{Is used to compute exactly one measurement for each network
computed in the temporal sequence of networks. Can contain the name of any
function which reduces a matrix to just one value or one of the following
character strings: \code{"bipolarization"} (for calculating the modularity of
different two-cluster solutions and choosing the highest one);
\code{"multipolarization"} (for calculating the modularity of different
community detection measures as implemented in \pkg{igraph}; or \code{"louvain"}
(for detecting communities using the fast Louvain algorithm and computing
modularity based on this result).}

\item{timewindow}{Same as in \link{dna_network}.}

\item{windowsize}{Same as in \link{dna_network}.}

\item{facet}{Which value from the dna database should be used to subset the
networks. Can be \code{"Authors"} for document author, \code{"Sources"} for
document source, \code{"Sections"} for documents which contain a certain
section or \code{"Types"} to subset document types.}

\item{facetValues}{Which values should be used to facet calculation of the
networks. Always contains the value "all" for comparison. Use e.g.
excludeTypes to exclude documents from comparison.}

\item{parallel}{Use parallel processing for the computation of time series
measures? By default, this is switched off (\code{"no"}), meaning that only
one CPU node is used to compute the results. Other possible values are
\code{"FORK"} and \code{"PSOCK"}. The FORK option uses the \code{mclapply}
function under the hood. This works well on Unix systems with multiple
cores but not on Windows. If \code{"FORK"} is used, the \code{cl} argument is
ignored. The \code{"PSOCK"} option uses the \code{parSapply} function under
the hood. This works on any kind of system including Windows but is
slightly slower than forking. In either case, the number of cores can be
provided using the \code{ncpus} argument.}

\item{ncpus}{Number of CPU cores to use for parallel processing (if switched
on through the \code{parallel} argument).}

\item{cl}{An optional \code{cluster} object for PSOCK parallel processing.
If no \code{cluster} object is provided (default behaviour), a cluster is
created internally on the fly. If a \code{cluster} object is provided,
the \code{ncpus} argument is ignored because the \code{cluster} object
already contains the number of cores. It can be useful to supply a
\code{cluster} object for example for use with the MPI protocol on a HPC
cluster. Note that the \code{cl} argument is ignored if
\code{parallel = "FORK"}.}

\item{verbose}{Display messages if \code{TRUE} or \code{1}. Also displays
details about network construction when \code{2}.}

\item{...}{Additional arguments passed to \link{dna_network}.}
}
\description{
Computes a measure for each network in a temporal sequence of networks.
}
\details{
This function serves as a convenience wrapper to calculate a measure for each
network in a temporal sequence of networks. The standard is to calculate the
modularity the best two-community solution (i.e., bipolarization) of the
network for each time window. The function can also be used to split
your data (facet) and calculate networks for each facet type and return the
statistics by facet.
}
\examples{
\dontrun{
library("ggplot2")
dna_init()
conn <- dna_connection(dna_sample())

tW <- dna_timeWindow(connection = conn,
                     timewindow = "days",
                     windowsize = 10,
                     facet = "Authors",
                     facetValues = c("Bluestein, Joel",
                                     "Voinovich, George"),
                     method = "bipolarization",
                     verbose = TRUE)

dna_plotTimeWindow(tW,
                   facetValues = c("Bluestein, Joel",
                                   "Voinovich, George",
                                   "all"),
                   rows = 2)

mp <- dna_timeWindow(connection = conn,
                     timewindow = "days",
                     windowsize = 15,
                     method = "multipolarization",
                     duplicates = "document",
                     parallel = "PSOCK",
                     ncpus = 3)

dna_plotTimeWindow(mp, include.y = c(-1, 1)) + theme_bw()
}

}
\author{
Philip Leifeld, Johannes B. Gruber
}
