% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_toIgraph}
\alias{dna_toIgraph}
\title{Convert DNA networks to igraph objects}
\usage{
dna_toIgraph(x, weighted = TRUE)
}
\arguments{
\item{x}{A dna_network (one- or two-mode).}

\item{weighted}{Logical. Should edge weights be used to create a weighted
graph from the dna_network object.}
}
\description{
This function can convert objects of class 'dna_network_onemode' or
'dna_network_twomode' to igraph objects.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())
nw <- dna_network(conn, networkType = "onemode")
graph <- dna_toIgraph(nw)
}
}
