% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{dna_toREM}
\alias{dna_toREM}
\title{Convert DNA networks to eventSequence objects}
\usage{
dna_toREM(x, variable = "organization", ...)
}
\arguments{
\item{x}{A \code{dna_connection} object created by the \code{dna_connection}
function or a \code{dna_eventlist} object created with \code{dna_network}
(setting \code{networkType = "eventlist"}).}

\item{variable}{The first variable for network construction(see
\link{dna_network}). The second one defaults to "concept" but can be
provided via ... if necessary (see \code{variable2} in
\code{dna_connection}).}

\item{...}{Additional arguments passed to \link{dna_network} and
\link[rem]{eventSequence}.}
}
\value{
A data.frame containing an event sequence for relational event
models.
}
\description{
This function can produce eventSequence objects (see
\link[rem]{eventSequence}) used in the package rem from DNA connections.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample())

### Convert dna_connection to eventSequence
eventSequence <- dna_toREM(conn)

### Convert network object to eventSequence
nw <- dna_network(conn, networkType = "eventlist")
eventSequence <- dna_toREM(nw)

### Pass on arguments to dna_network
eventSequence2 <- dna_toREM(conn, duplicates = "document")

### Pass on arguments to eventSequence
eventSequence3 <- dna_toREM(conn, excludeTypeOfDay = "Wednesday")
}
}
