% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_connection}
\alias{print.dna_connection}
\title{Print the summary of a \code{dna_connection} object}
\usage{
\method{print}{dna_connection}(x, ...)
}
\arguments{
\item{x}{A \code{dna_connection} object.}

\item{...}{Further options (currently not used).}
}
\description{
Show details of a \code{dna_connection} object.
}
\details{
Print the number of documents and statements to the console after
establishing a DNA connection.
}
\examples{
\dontrun{
dna_init()
conn <- dna_connection(dna_sample(), verbose = FALSE)
conn
}
}
