% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{print.dna_documents}
\alias{print.dna_documents}
\title{Print the summary of a \code{dna_documents} object}
\usage{
\method{print}{dna_documents}(x, truncate = 20, ...)
}
\arguments{
\item{x}{A \code{dna_connection} object.}

\item{truncate}{Number of characters to which character columns in the data
frame should be truncated.}

\item{...}{Further options (currently not used).}
}
\description{
Show details of a \code{dna_documents} object.
}
\details{
Print a data frame returned by \link{dna_getDocuments}. The only difference
between this print method and the default print method for data frames is
that the \code{text} column and other columns containing character strings
are truncated for better readability on screen.
}
