% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDNA.R
\name{trim}
\alias{trim}
\title{Truncate labels}
\usage{
trim(x, n)
}
\arguments{
\item{x}{A character string}

\item{n}{May number of character to truncate to}
}
\description{
Internal function, used to truncate labels
}
\author{
Philip Leifeld, Johannes B. Gruber
}
